@@ -2,15 +2,26 @@
   "name": "chop_chow_sd",
   "version": "0.1.0",
   "engines": {
-    "node": "18.x"
+    "node": "20.x"
   },
   "private": true,
   "dependencies": {
     "@aws-sdk/client-cognito-identity": "^3.41.0",
-    "@aws-sdk/client-s3": "^3.41.0",
+    "@aws-sdk/client-s3": "^3.616.0",
     "@aws-sdk/credential-provider-cognito-identity": "^3.41.0",
+    "@distube/ytdl-core": "^4.14.4",
     "@emotion/react": "^11.9.3",
     "@emotion/styled": "^11.9.3",
+    "@fortawesome/fontawesome-svg-core": "^6.5.2",
+    "@fortawesome/free-regular-svg-icons": "^6.5.2",
+    "@fortawesome/free-solid-svg-icons": "^6.5.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@hotjar/browser": "^1.0.9",
+    "@lexical/history": "^0.17.0",
+    "@lexical/html": "^0.17.0",
+    "@lexical/link": "^0.17.0",
+    "@lexical/list": "^0.17.0",
+    "@lexical/react": "^0.17.0",
     "@mui/icons-material": "^5.8.4",
     "@mui/lab": "^5.0.0-alpha.86",
     "@mui/material": "^5.8.4",
@@ -19,17 +30,29 @@
     "@next/codemod": "^14.0.1",
     "@next/third-parties": "^14.1.2",
     "@react-google-maps/api": "^2.19.2",
-    "@react-oauth/google": "^0.8.0",
+    "@react-oauth/google": "^0.12.1",
     "@redux-devtools/extension": "3.0.0",
+    "@reduxjs/toolkit": "^2.2.7",
+    "@sasmeee/igdl": "^1.0.0",
+    "autoprefixer": "^10.4.19",
     "axios": "^0.27.2",
     "axios-auth-refresh": "^3.3.6",
+    "bcrypt": "^5.1.1",
+    "bcryptjs-react": "^2.4.6",
     "body-scroll-lock": "^4.0.0-beta.0",
+    "card-validator": "^10.0.0",
     "chart.js": "^4.0.1",
+    "crypto-hash": "^3.0.0",
+    "crypto-js": "^4.2.0",
+    "css-loader": "^7.1.2",
     "dayjs": "^1.11.6",
     "ejs": "^3.1.6",
+    "fluent-ffmpeg": "^2.1.3",
     "iso-country-currency": "^0.7.2",
     "jquery": "^3.5.1",
-    "jwt-decode": "^3.1.2",
+    "jszip": "^3.10.1",
+    "jwt-decode": "^4.0.0",
+    "lexical": "^0.17.0",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
     "next": "^14.1.2",
@@ -38,6 +61,8 @@
     "os": "^0.1.2",
     "passport-facebook": "^3.0.0",
     "path": "^0.12.7",
+    "postcss": "8.0.0",
+    "postcss-loader": "^8.1.1",
     "pure-react-carousel": "^1.25.1",
     "react": "18.2.0",
     "react-bootstrap": "^1.0.0-beta.8",
@@ -45,6 +70,7 @@
     "react-chartjs-2": "^5.0.1",
     "react-dom": "18.2.0",
     "react-facebook-login": "^4.1.1",
+    "react-fontawesome": "^1.7.1",
     "react-geolocated": "^4.1.2",
     "react-google-autocomplete": "^2.7.3",
     "react-google-places-autocomplete": "^4.0.1",
@@ -64,6 +90,7 @@
     "react-simple-star-rating": "^5.1.7",
     "react-slick": "^0.29.0",
     "react-spinners": "^0.13.8",
+    "react-switch": "^7.0.0",
     "react-tabs-redux": "^4.0.0",
     "react-to-print": "^2.14.4",
     "react-toast": "^1.0.3",
@@ -77,7 +104,9 @@
     "slick-carousel": "^1.8.1",
     "source-map-explorer": "^2.5.2",
     "styled-components": "^5.3.6",
-    "yarn": "^1.22.19"
+    "uuid": "^10.0.0",
+    "yarn": "^1.22.19",
+    "zod": "^3.23.8"
   },
   "scripts": {
     "analyze": "source-map-explorer 'build/static/js/*.js'",
@@ -94,8 +123,10 @@
   ],
   "devDependencies": {
     "@babel/core": "^7.5.5",
+    "@types/node": "20.11.24",
     "local-ssl-proxy": "^2.0.5",
-    "npm-run-all": "^4.1.3"
+    "npm-run-all": "^4.1.3",
+    "typescript": "5.3.3"
   },
   "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
   "main": "index.js",
