@@ -1,4 +1,8 @@
 import Head from "next/head";
+import Header, { Header2 } from "../src/components/Header/Header";
+import OTP from "../src/components/OTP";
+import Resetpassword from "../src/components/resetpassword"
+
 const ResetPassword = () => {
     return <div>
         <Head>
@@ -6,7 +10,13 @@ const ResetPassword = () => {
             <meta key="title" name="viewport" content="initial-scale=1.0, width=device-width" />
         </Head>
 
-        Welcome to Next.js!</div>
+        <div>
+            <Header/>
+            
+        <Resetpassword/>
+        
+        </div>
+        </div>
 
 }
 
