@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import React from "react";
 import Head from "next/head";
 import Header, { Header2 } from "../Header/Header";
 import TimeBar from "../GroceryPage/CartPage/Cart/TimeBar";
@@ -7,34 +7,75 @@ import CartItem from "../GroceryPage/CartPage/CartItem/Index";
 import Footer from "../Footer/Footer";
 import CartContext from "../../../pages/store/cart-context";
 import SideNav from "../Header/sidenav";
-import { useMobileMedia } from '../../customhooks/useResponsive';
+import { useMobileMedia } from "../../customhooks/useResponsive";
+import { connect, useDispatch, useSelector } from "react-redux";
+import {
+  addToCart,
+  deleteFromCart,
+  EmptyCart,
+  FetchCart,
+  removeFromCart,
+} from "../../actions/Cart";
+import { useRouter } from "next/router";
+import { v4 as uuidv4 } from "uuid";
+import { useEffect } from "react";
 
-function Index(props) {
-
+function Cart(props) {
   const mobileScreen = useMobileMedia();
+  const { cartItems: items } = useSelector((state) => {
+    return state.Cart;
+  });
 
-  const cartCtx = useContext(CartContext);
+  const dispatch = useDispatch();
+  const router = useRouter();
 
-  const totalAmount = `$${cartCtx.totalAmount.toFixed(2)}`;
+  useEffect(() => {
+    console.log("getting cart");
+    dispatch(FetchCart());
+  }, []);
 
-  const hasItems = cartCtx.items.length > 0;
+  const hasItems = items?.length > 0;
 
-  const cartItemRemoveHandler = (id) => {
-    cartCtx.removeItem(id);
+  const totalQuantity = `${items?.reduce((a, c) => a + c.amount, 0)} items`;
+  const totalPrice = items
+    ?.reduce((a, c) => a + c.price * c.amount, 0)
+    .toFixed(2);
+
+  const Checkout = () => {
+    router.push(`/checkout/`);
   };
 
-  const cartItemAddHandler = (item) => {
-    cartCtx.addItem({ ...item, amount: 1 });
+  const AddToCart = (item) => {
+    const payload = {
+      itemName: item.name,
+      item_image: item.picture,
+      item_price: item.price,
+      itemId: item.id,
+      userId: item.userId || "",
+      storeName: item.store,
+      currency: item.currency,
+      quantity: item.amount,
+      storeId: item.storeId,
+      item_type: item.item_type ? item.item_type : "",
+    };
+    try {
+      dispatch(addToCart(payload));
+    } catch (error) {
+      console.log(error);
+    }
   };
 
-  const cartItemClearHandler = (id) => {
-    cartCtx.clearItem(id);
+  const RemoveFromCart = (product) => {
+    dispatch(removeFromCart(product.id));
   };
 
-  const cartCartClearHandler = (id) => {
-    cartCtx.clearCart(id);
+  const CloseCart = () => {
+    router.push("/marketplace");
   };
 
+  const DeleteFromCart = (product) => {
+    dispatch(deleteFromCart(product.id));
+  };
   return (
     <div className={indexStyles.cartBackground}>
       <Head>
@@ -47,8 +88,8 @@ function Index(props) {
       </Head>
       <Header route="groceryList" />
       <Header2 />
-      <SideNav/> 
-      <TimeBar />
+      <SideNav />
+      {/* <TimeBar />  */}
       <div className={indexStyles.cartBody}>
         <div className={indexStyles.cartMainBody}>
           <div className={indexStyles.arrowBack}>
@@ -57,40 +98,45 @@ function Index(props) {
                 id="backArr"
                 src="/assets/grocery_list/backArr.svg"
                 alt="arrowDown"
-                onClick={props.closeCart}
+                onClick={CloseCart}
               />
               <label onClick={props.closeCart}>Back</label>
             </div>
-            <h1>CART</h1>
+            <h2 style={{ fontWeight: "1000px", fontSize: "3.5rem" }}>Cart</h2>
           </div>
-          {!mobileScreen ? <div className={indexStyles.cartHeader}>
-            <div className={indexStyles.cartHeaderRow1}>
-              <label>Product</label>
+          {!mobileScreen ? (
+            <div className={indexStyles.cartHeader}>
+              <div className={indexStyles.cartHeaderRow1}>
+                <label>Product</label>
+              </div>
+              <div className={indexStyles.cartHeaderRow2}>
+                <label>Quantity</label>
+              </div>
+              <div className={indexStyles.cartHeaderRow3}>
+                <label>Store</label>
+                <label>Price</label>
+                <label>Subtotal</label>
+                <label>Action</label>
+              </div>
             </div>
-            <div className={indexStyles.cartHeaderRow2}>
-              <label>Quantity</label>
-            </div>
-            <div className={indexStyles.cartHeaderRow3}>
-              <label>Store</label>
-              <label>Price</label>
-              <label>Subtotal</label>
-              <label>Action</label>
-            </div>
-          </div> : "Items"}
-          {cartCtx.items.map((item) => (
+          ) : (
+            "Items"
+          )}
+          {items?.map((item) => (
             <CartItem
-              key={item.id}
-              id={item.id}
+              key={item.itemId}
+              id={item.itemId}
               name={item.name}
               amount={item.amount}
-              picture={item.picture}
+              picture={item.image}
               price={item.price}
-              store={item.store}
-              totalAmount={totalAmount}
-              pickUpTime={item.pickUpTime}
-              onRemove={cartItemRemoveHandler.bind(null, item.id)}
-              onAdd={cartItemAddHandler.bind(null, item)}
-              onClearItem={cartItemClearHandler.bind(null, item.id)}
+              store={item.storeName}
+              totalAmount={item.totalAmount}
+              storeId={item.storeId}
+              userId={item.userId}
+              onAdd={AddToCart}
+              onRemove={RemoveFromCart}
+              onDelete={DeleteFromCart}
             />
           ))}
           <div className={indexStyles.couponPrice}>
@@ -106,18 +152,26 @@ function Index(props) {
             <div className={indexStyles.priceRow}>
               <div className={indexStyles.mainPrice}>
                 <label>Total Price</label>
-                <label>{totalAmount}</label>
+
+                <label>
+                  {`${Number(totalPrice) == "NAN" ? 0 : Number(totalPrice)} `}{" "}
+                </label>
               </div>
               <div className={indexStyles.checkoutButton}>
                 <div className={indexStyles.whiteButton}>
                   {hasItems ? (
-                    <label onClick={cartCartClearHandler}>Clear Cart</label>
+                    <label onClick={() => dispatch(EmptyCart())}>
+                      Clear Cart
+                    </label>
                   ) : (
-                    <label onClick={props.closeCart}>Add Items</label>
+                    <label onClick={CloseCart}>Add Items</label>
                   )}
                 </div>
                 {hasItems && (
-                  <div className={indexStyles.greenButton}>
+                  <div
+                    className={indexStyles.greenButton}
+                    onClick={() => Checkout()}
+                  >
                     <label>Check Out</label>
                   </div>
                 )}
@@ -131,4 +185,4 @@ function Index(props) {
   );
 }
 
-export default Index;
+export default Cart;
