@@ -2,6 +2,7 @@ import styles from "./suggesteddescription.module.css";
 import axios from "../../util/Api";
 import { CloseFillIcon, FillterIcon } from "../../components/icons";
 import { IoMdCloseCircle } from "react-icons/io";
+import { useRouter } from "next/router";
 
 export const SuggestedDescription = ({
   descriptions,
@@ -9,6 +10,8 @@ export const SuggestedDescription = ({
   deleteDescription,
   status,
   setStatus,
+  filteredDescription,
+  handleFilteredDescription
 }) => {
   const handleStatus = () => {
     if (status === "all") {
@@ -23,6 +26,7 @@ export const SuggestedDescription = ({
       setStatus("all");
     }
   };
+  const router = useRouter()
   return (
     <div className={styles.container}>
       <table
@@ -30,19 +34,59 @@ export const SuggestedDescription = ({
       >
         <thead>
           <tr className={styles.th1}>
-            <td className={styles.td}>Name</td>
+            <td className={styles.td} onClick={() => {
+              handleFilteredDescription(
+                "description_key",
+                Number(filteredDescription?.description_key) === 1
+                  ? -1
+                  : 1
+              );
+            }}>Name</td>
             <td className={styles.td}>
               <p>
-                Formatted string <FillterIcon style={styles.FillterIcon} />
+                Formatted string
               </p>
             </td>
             <td className={styles.td}>
-              <p style={{ cursor: "pointer" }} onClick={handleStatus}>
+              <p style={{ cursor: "pointer" }} onClick={() => {
+                let item_status = "Public";
+                if (router?.query?.item_status === "Public") {
+                  item_status = "Draft";
+                } else if (
+                  router?.query?.item_status === "Draft"
+                ) {
+                  item_status = "Pending";
+                } else if (
+                  router?.query?.item_status === "Pending"
+                ) {
+                  item_status = "Rejected";
+                } else if (
+                  router?.query?.item_status === "Rejected"
+                ) {
+                  item_status = "Public";
+                }
+                console.log(item_status, 'itemstatus', router?.query, router?.query?.item_status === 'Public')
+                router.push({
+                  pathname: router.pathname,
+                  query: {
+                    ...router.query,
+                    item_status,
+                  },
+                });
+                handleFilteredDescription("status", item_status);
+              }}>
                 Status <FillterIcon style={styles.FillterIcon} />
               </p>
             </td>
             <td className={styles.td}>
-              <p>
+              <p onClick={() => {
+                handleFilteredDescription(
+                  "createdAt",
+                  Number(filteredDescription?.createdAt) === 1
+                    ? -1
+                    : 1
+                );
+              }}>
                 Date created <FillterIcon style={styles.FillterIcon} />
               </p>
             </td>
@@ -62,10 +106,10 @@ export const SuggestedDescription = ({
                     element.status === "Public"
                       ? styles.statusText
                       : element.status === "Pending"
-                      ? styles.statusText2
-                      : element.status === "Rejected"
-                      ? styles.rejected
-                      : styles.statusText2
+                        ? styles.statusText2
+                        : element.status === "Rejected"
+                          ? styles.rejected
+                          : styles.statusText2
                   }
                   style={{ textTransform: "capitalize" }}
                 >
