@@ -12,6 +12,8 @@ import { useState } from "react";
 import Select from "react-select";
 
 import ArrowDropUp from "@mui/icons-material/ArrowDropUp";
+import { ReceivedModal } from "../modal/received-rejection";
+import { useSelector } from "react-redux";
 function SuggestedMealRow(props) {
   const [show, setShowState] = useState(false);
   const months = [
@@ -38,34 +40,52 @@ function SuggestedMealRow(props) {
   }
   const [selectedAction, setSelectedAction] = useState(null);
   const [showPopup, setShowPopup] = useState(false);
+  const [showRejection, setShowRejection] = useState(false);
 
   const handleClickPopup = () => {
     setShowPopup(true);
   };
+  const selectedUserType = useSelector((state) => state.userType.selectedUserType);
 
   const options = [
     { value: "default", label: "Select..." },
     {
       value: "sendForReview",
       label: "Send for review",
-      isDisabled: suggestion.item_status[0]?.status !== "Draft",
+      // isDisabled:
+      //   suggestion.item_status[0]?.status !== "Draft" &&
+      //   suggestion.item_status[0]?.status !== "Rejected",
     },
     {
       value: "availableInInventory",
       label: "Available in Inventory",
-      isDisabled: suggestion.item_available === false || suggestion.item_available === undefined,
+      // isDisabled:
+      //   suggestion.item_available === false ||
+      //   suggestion.item_available === undefined,
     },
     {
       value: "sendToInventory",
       label: "Send to Inventory",
-      isDisabled: !(suggestion.item_status[0]?.status === "Public"),
+      // isDisabled: !(suggestion.item_status[0]?.status === "Public"),
     },
+    {
+      value: "Draft",
+      label: "Draft",
+      // isDisabled:
+      //   suggestion.item_status[0]?.status === "Draft",
+    }
+
   ];
-  const handleSelectChange = (selectedOption) => {
+  const handleSelectChange = async (selectedOption) => {
+    console.log("Selected option:", selectedOption);
     setSelectedAction(selectedOption);
 
     if (selectedOption.value === "sendForReview") {
       props.handleSendForReview(suggestion._id, "Pending");
+    } else if (selectedOption.value === "Draft") {
+      props.handleStatusType("Draft", suggestion._id)
+      console.log(suggestion, 'idd')
+
     } else if (
       selectedOption.value === "sendToInventory" &&
       suggestion.item_status[0]?.status === "Public"
@@ -73,14 +93,15 @@ function SuggestedMealRow(props) {
       props.toggleTransferToInventory(suggestion);
     }
   };
+
+
   const customStyles = {
     control: (provided) => ({
       ...provided,
-      width: "100%", // Set your desired fixed width
+      width: "100%",
     }),
   };
-
-  console.log(suggestion, "hells");
+  console.log(suggestion, "suggested");
   return (
     <div key={suggestion._id} className={styles.request_tr_div}>
       <table
@@ -92,7 +113,7 @@ function SuggestedMealRow(props) {
         <tbody>
           <tr
             className={
-              props.auth.authUser.user_type === "supplier"
+              selectedUserType === "supplier"
                 ? styles.meal_tr
                 : styles.meal_tr1
             }
@@ -100,98 +121,106 @@ function SuggestedMealRow(props) {
             <td className={styles.td_name}>
               <p
                 onClick={
-                  props.auth.authUser.user_type === "admin"
+                  selectedUserType === "admin"
                     ? () => props.toggleOpenMeal(suggestion)
                     : props.searchType === "Item"
-                    ? () => props.openMealDetailsModal(suggestion)
-                    : () => props.openDetailsModal(suggestion)
+                      ? () => props.openMealDetailsModal(suggestion)
+                      : () => props.openDetailsModal(suggestion)
+
                 }
               >
-                {props.auth.authUser.user_type === "supplier" &&
-                props.searchType === "Item"
+                {selectedUserType === "supplier" &&
+                  props.searchType === "Item"
                   ? suggestion.item_name.length > 25
                     ? suggestion.item_name.slice(0, 25) + "..."
                     : suggestion.item_name
                   : props.searchType === "Item"
-                  ? suggestion.item_name
-                  : suggestion?.category_name}
+                    ? suggestion.item_name
+                    : suggestion?.category_name}
               </p>
             </td>
             <td className={styles.td_name}>
               <p
                 onClick={
-                  props.auth.authUser.user_type === "admin"
+                  selectedUserType === "admin"
                     ? () => props.toggleOpenMeal(suggestion)
                     : props.searchType === "Item"
-                    ? () => props.openMealDetailsModal(suggestion)
-                    : () => props.openDetailsModal(suggestion)
+                      ? () => props.openMealDetailsModal(suggestion)
+                      : () => props.openDetailsModal(suggestion)
                 }
                 className={styles.hide}
               >
                 {props.searchType === "Item"
                   ? suggestion.item_type
                   : props.searchType === "Product"
-                  ? suggestion.item_type
-                  : suggestion?.category_name}
+                    ? suggestion.item_type
+                    : suggestion?.category_name}
               </p>
             </td>
-            <td className={styles.td_name}>
+            <td
+              className={styles.td_name}
+              onMouseEnter={() => setShowRejection(true)}
+              onMouseLeave={() => setShowRejection(false)}
+            >
               <p
                 onClick={
-                  props.auth.authUser.user_type === "admin"
+                  selectedUserType === "admin"
                     ? () => props.toggleOpenMeal(suggestion)
                     : props.searchType === "Item"
-                    ? () => props.openMealDetailsModal(suggestion)
-                    : () => props.openDetailsModal(suggestion)
+                      ? () => props.openMealDetailsModal(suggestion)
+                      : () => props.openDetailsModal(suggestion)
                 }
                 className={
                   status +
                   " " +
                   (suggestion.item_status[0]?.status === "Draft" ||
-                  suggestion.item_status[0]?.status === "Pending"
+                    suggestion.item_status[0]?.status === "Pending"
                     ? pending
                     : suggestion.item_status[0]?.status === "Public"
-                    ? approve
-                    : suggestion.item_status[0]?.status === "Rejected"
-                    ? rejected
-                    : "")
+                      ? approve
+                      : suggestion.item_status[0]?.status === "Rejected"
+                        ? rejected
+                        : "")
                 }
               >
                 {props.searchType !== "Category"
                   ? suggestion.item_status[0]?.status
                   : suggestion.publicly_available}
-                {/* {suggestion.item_status[0].status} */}
               </p>
+
+              {suggestion.item_status[0]?.status === "Rejected" &&
+                suggestion._id &&
+                showRejection && <ReceivedModal suggestion={suggestion} />}
             </td>
             <td className={styles.td_cat}>
               <p
                 onClick={
-                  props.auth.authUser.user_type === "admin"
+                  selectedUserType === "admin"
                     ? () => props.toggleOpenMeal(suggestion)
                     : props.searchType === "Item"
-                    ? () => props.openMealDetailsModal(suggestion)
-                    : () => props.openDetailsModal(suggestion)
+                      ? () => props.openMealDetailsModal(suggestion)
+                      : () => props.openDetailsModal(suggestion)
                 }
                 className={styles.hideData}
               >
                 {props.searchType === "Item"
                   ? suggestion.item_categories.length > 0 &&
-                    suggestion.item_categories
-                      .map((ele) => ele?.category_name)
-                      .join(", ")
+                  suggestion.item_categories
+                    .map((ele) => ele?.category_name)
+                    .join(", ")
                   : suggestion.product_categories &&
-                    suggestion.product_categories.length > 0 &&
-                    suggestion.product_categories[0]}
+                  suggestion.product_categories.length > 0 &&
+                  suggestion.product_categories[0]}
               </p>
             </td>
             <td className={styles.td_cat}>
               <p
                 onClick={
-                  props.auth.authUser.user_type === "admin"
+                  selectedUserType === "admin"
                     ? () => props.toggleOpenMeal(suggestion)
                     : props.searchType === "Item"
-                    ? () => props.openMealDetailsModal(suggestion)
-                    : () => props.openDetailsModal(suggestion)
+                      ? () => props.openMealDetailsModal(suggestion)
+                      : () => props.openDetailsModal(suggestion)
                 }
                 className={styles.hideData}
               >
@@ -201,29 +230,31 @@ function SuggestedMealRow(props) {
             </td>
             <td className={styles.td_name}>
               <div className={styles.actions_con}>
-                {props.auth.authUser.user_type !== "admin" && (
+                {selectedUserType !== "admin" && (
                   <div className={styles.selectContainer}>
                     <Select
                       styles={customStyles}
                       defaultValue={selectedAction}
                       onChange={handleSelectChange}
                       options={options.filter((option) => {
-                        if (suggestion.item_status[0]?.status === "Draft") {
+                        const status = suggestion.item_status[0]?.status;
+                        const hasPrice = !!suggestion.item_price;
+                        if (status === "Draft") {
                           return option.value === "sendForReview";
-                        } else if (
-                          suggestion.item_status[0]?.status === "Public" &&
-                          suggestion.item_price
-                        ) {
-                          return option.value === "availableInInventory";
                         }
-                        return true;
+                        if (status === "Public") {
+                          if (option.value === "Draft") return true;
+                          if (option.value === "availableInInventory") return hasPrice;
+                          if (option.value === "sendToInventory") return !hasPrice;
+                        }
+                        return false;
                       })}
+
                     />
                   </div>
                 )}
 
                 <i
-                  className={styles.hideData}
                   onClick={() => {
                     // props.deleteSuggestion(suggestion._id)
 
@@ -232,15 +263,7 @@ function SuggestedMealRow(props) {
                 >
                   <CloseFillIcon style={actionIcon} />
                 </i>
-                {show ? (
-                  <i onClick={showDropDown} className={styles.showData}>
-                    <ArrowDropUp className={styles.arrowDown} />
-                  </i>
-                ) : (
-                  <i onClick={showDropDown} className={styles.showData}>
-                    <ArrowDropDownIcon className={styles.arrowDown} />
-                  </i>
-                )}
+
                 {showPopup && (
                   <div className={styles.addpublicMeal_container}>
                     <div className={styles.popup}>
