@@ -1,65 +1,84 @@
 import CancelIcon from "@mui/icons-material/Cancel";
 import styles from "./transferToInventory.module.css";
 import styled from "styled-components";
-import Switch from "@mui/material/Switch";
+import Switch from "react-switch";
 import { useEffect, useState } from "react";
 import axios from "../../util/Api";
 import ArrowDropDownIcon from "@mui/icons-material/ArrowDropDown";
 import Image from "next/image";
 import { AiFillInfoCircle } from "react-icons/ai";
+import { useMediaQuery } from "../../hooks/usemediaquery";
+import Select from "react-select";
+import { toast } from "react-toastify";
 
+// const CustomSwitch = ({ checked, onChange }) => {
+//   return (
+//     <SwitchContainer onClick={() => onChange(!checked)}>
+//       <SwitchSlider checked={checked}>
+//         <SwitchBackground />
+//         <SwitchLabel checked={checked}>{checked ? "Yes" : "No"}</SwitchLabel>
+//       </SwitchSlider>
+//     </SwitchContainer>
+//   );
+// };
 const CustomSwitch = ({ checked, onChange }) => {
+  const handleClick = () => {
+    const newChecked = !checked;
+    onChange(newChecked);
+  };
+
   return (
-    <SwitchContainer onClick={() => onChange(!checked)}>
-      <SwitchSlider checked={checked} />
-      <SwitchLabel>{checked ? "Yes" : "No"}</SwitchLabel>
-    </SwitchContainer>
+    <div className={styles.switch} onClick={handleClick}>
+      <div className={styles.flexed}>
+        <p style={{ fontSize: "12px", marginLeft: "2px" }}>Yes</p>
+        <p style={{ fontSize: "12px", marginLeft: "2px" }}>No</p>
+      </div>
+      <div className={checked ? styles.roundSwitch : styles.roundSwitch2}>
+        <p>{checked ? "Yes" : "No"}</p>
+      </div>
+    </div>
   );
 };
 
 const SwitchContainer = styled.div`
-  display: inline-block;
   position: relative;
-  width: 70px;
-  height: 30px;
-  cursor: pointer;
+  display: inline-block;
+  width: 60px;
+  height: 34px;
+  background-color: "#FFF"
+  box-shadow: "0px 4px 5px 0px #0000001A",
+
 `;
 
 const SwitchSlider = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: ${(props) => (props.checked ? "#04D505" : "#949494")};
-  border-radius: 20px;
-  transition: background-color 0.3s;
-  pointer-events: none;
-  &::before {
-    content: "";
-    position: absolute;
-    width: 30px;
-    height: 30px;
-    left: ${(props) => (props.checked ? "calc(100% - 30px)" : "0")};
-    top: 50%;
-    transform: translateY(-50%);
-    background-color: white;
-    border-radius: 50%;
-    transition: left 0.3s;
-  }
+  position: relative;
+  cursor: pointer;
+  width: 60px;
+  height: 34px;
+  background-color: ${(props) => (props.checked ? "#fff" : "#4CAF50")};
+  border-radius: 34px;
 `;
 
-const SwitchLabel = styled("span")`
+const SwitchBackground = styled.div`
+  position: absolute;
+  width: ${(props) => (props.checked ? "100%" : "0%")};
+  height: 100%;
+  background-color: #4caf50;
+  border-radius: 34px;
+  transition: width 0.2s ease;
+`;
+
+const SwitchLabel = styled.span`
+  color: black;
   position: absolute;
   top: 50%;
-  left: 50%;
+  left: ${(props) => (props.checked ? "5px" : "calc(100% - 25px)")};
   transform: translate(-50%, -50%);
-  color: white;
-  font-size: 14px;
+  transition: left 0.2s ease;
 `;
 
 export default function TransferToInventory(props) {
-  console.log(props.checked, "pppp");
+  const matches = useMediaQuery("(min-width: 700px)");
   const [restockOption, setRestockOption] = useState();
   const [restockTime, setRestockTime] = useState("1 day");
   const [message, setMessage] = useState({
@@ -79,28 +98,6 @@ export default function TransferToInventory(props) {
   const { reloadData } = props;
 
   const { ingredientsAvailable, item_type, in_stock } = formState;
-  console.log(props, "pp");
-  // useEffect(() => {
-  //   console.log( props.meal.ingredeints_in_item, 'propsss');
-  //   if (props.meal.item_type === "Meal" && props.meal.ingredeints_in_item) {
-  //     const ingredientsAvailable = props.meal.ingredeints_in_item;
-
-  //     const newIngredients = ingredientsAvailable?.map((ingredient) => ({
-  //       name: ingredient.item_name,
-  //       quantity: ingredient.quantity,
-  //       set_price: "",
-  //       product_available: true,
-  //     }));
-
-  //     console.log(newIngredients, 'new');
-
-  //     setFormState((prevState) => ({
-  //       ...prevState,
-  //       ingredientsAvailable: newIngredients,
-  //       item_type: props.type,
-  //     }));
-  //   }
-  // }, [props.meal.item_type, props.meal.ingredeints_in_item]);
 
   function handleChange(e) {
     const { name, value } = e.target;
@@ -121,18 +118,7 @@ export default function TransferToInventory(props) {
     } else {
       setFormState({ ...formState, [name]: value });
     }
-
-    console.log(formState);
   }
-
-  // useEffect(() => {
-  //   if (props?.ingredeints_in_item) {
-  //     setFormState({
-  //       ...formState,
-  //       ingredientsAvailable: props.ingredeints_in_item,
-  //     });
-  //   }
-  // }, [props?.ingredeints_in_item]);
 
   function handleInStockChange(value) {
     setFormState({ ...formState, in_stock: value });
@@ -140,10 +126,8 @@ export default function TransferToInventory(props) {
 
   function handleIngredientChange(e, id, key) {
     const { value } = e.target;
-    console.log(e, id, key);
     let ingredientsAvailable = formState.ingredientsAvailable;
     ingredientsAvailable[id][key] = value;
-    console.log(ingredientsAvailable);
     setFormState({
       ...formState,
       ["ingredientsAvailable"]: ingredientsAvailable,
@@ -151,21 +135,29 @@ export default function TransferToInventory(props) {
   }
 
   function handleIngredientAvailabilityChange(value, id, key) {
-    console.log(value);
-    let ingredientsAvailable = formState.ingredientsAvailable;
-    ingredientsAvailable[id][key] = value;
+    // Create a shallow copy of the ingredientsAvailable array
+    const updatedIngredientsAvailable = [...formState.ingredientsAvailable];
+
+    // Update the availability of the ingredient at the specified index
+    updatedIngredientsAvailable[id] = {
+      ...updatedIngredientsAvailable[id],
+      [key]: value,
+    };
+
+    // Update the formState with the updated ingredientsAvailable array
     setFormState({
       ...formState,
-      ["ingredientsAvailable"]: ingredientsAvailable,
+      ingredientsAvailable: updatedIngredientsAvailable,
     });
   }
 
   function sendToInventory() {
     let fields = formState;
-
+    if (!fields["meal_price"] || fields["meal_price"] == 0) {
+      toast.error(`Meal price cannot be empty!`);
+    }
     delete fields.meal_type;
     fields["item"] = props.meal._id;
-    // fields["storeId"] = "63783f54088dda05688af4df";
     fields.ingredients = formState.ingredientsAvailable;
     delete formState.ingredientsAvailable;
     fields.user = JSON.parse(localStorage.getItem("user"))?._id ?? "";
@@ -173,9 +165,6 @@ export default function TransferToInventory(props) {
       .post("/inventory/create-inventory", fields)
       .then((response) => {
         if (response.status >= 200 && response.status < 300) {
-          // this.setState({ booleanOfDisplayOfDialogBoxConfirmation: true });
-          console.log(response);
-          console.log("Display Item submitted successfully");
           props.setTransferToInventoryState(false);
           setShow(true);
           setTimeout(() => {
@@ -215,7 +204,6 @@ export default function TransferToInventory(props) {
   function toggleRestockTimeOption() {
     setRestockOption(!restockOption);
   }
-  console.log(props, "ty");
 
   useEffect(() => {
     if (
@@ -250,7 +238,6 @@ export default function TransferToInventory(props) {
           "Content-Type": "application/json",
         },
       });
-      console.log(response.data.data, "one store");
       setAllStores(response.data.data);
       setCurrency(response.data.data);
     } catch (error) {
@@ -260,7 +247,15 @@ export default function TransferToInventory(props) {
   useEffect(() => {
     fetchOneUserStore();
   }, []);
-  console.log(formState.ingredientsAvailable, "avail");
+  if (in_stock) {
+  } else {
+  }
+
+  const storeOptions = allStores?.map((elem) => ({
+    value: elem?._id,
+    label: elem?.store_name,
+  }));
+
   return (
     <>
       {show && (
@@ -327,33 +322,43 @@ export default function TransferToInventory(props) {
                   borderBottom: "1px solid rgba(0, 0, 0, 0.2)",
                 }}
               >
-                <label htmlFor="mySelect" style={{ paddingBottom: "1rem" }}>
+                <label
+                  htmlFor="mySelect"
+                  style={{ paddingBottom: "1rem", fontSize: "14px" }}
+                >
                   Which store are you sending from?
                 </label>
-                <select
-                  style={{
-                    width: "25%",
-                    height: "34px",
-                    border: "1px solid #E6E6E6",
-                    outline: "none",
-                    borderRadius: "4px",
-                  }}
-                  onChange={(e) => {
-                    setFormState({
-                      ...formState,
-                      ["storeId"]: e.target.value,
-                    });
-                    setSelectedStore(e.target.value);
-                  }}
-                  name="storeId"
-                  value={formState.storeId}
-                >
-                  {allStores?.map((elem) => (
-                    <option key={elem?._id} value={elem?._id}>
-                      {elem?.store_name}
-                    </option>
-                  ))}
-                </select>
+
+                <div className={styles.dropdown}>
+                  {/* <button className={styles.dropdownButton}>Select Stores</button> */}
+                  <div>
+                    {allStores?.map((store) => (
+                      <label key={store._id} style={{ marginLeft: "1rem" }}>
+                        <input
+                          style={{ marginRight: ".5rem" }}
+                          type="checkbox"
+                          value={store._id}
+                          checked={
+                            formState.storeId?.includes(store._id) || false
+                          }
+                          onChange={(e) => {
+                            const storeIds = formState.storeId || [];
+                            const selectedValues = e.target.checked
+                              ? [...storeIds, e.target.value]
+                              : storeIds.filter((id) => id !== e.target.value);
+
+                            setFormState({
+                              ...formState,
+                              storeId: selectedValues,
+                            });
+                            setSelectedStore(selectedValues);
+                          }}
+                        />
+                        {store.store_name}
+                      </label>
+                    ))}
+                  </div>
+                </div>
               </div>
               <div className={styles.transToIn_meal_types}>
                 <p>Choose Meal Type</p>
@@ -484,13 +489,15 @@ export default function TransferToInventory(props) {
               </div>
 
               <div className={styles.transToIn_details_col3}>
-                <div>
-                  <h3>Are the ingredients available in your store</h3>
-                  <CustomSwitch
-                    checked={in_stock}
-                    onChange={handleInStockChange}
-                  />
-                </div>
+                {props?.meal?.item_type === "Meal" &&
+                  <div>
+                    <h3>Are the ingredients available in your store</h3>
+                    <CustomSwitch
+                      checked={in_stock}
+                      onChange={handleInStockChange}
+                    />
+                  </div>
+                }
                 <div>
                   <h3>Out of Stock? How long before restock</h3>
                   <div className={styles.select_container}>
@@ -547,23 +554,40 @@ export default function TransferToInventory(props) {
                         style={{ height: "2px" }}
                       >
                         <thead>
-                          <div
-                            className={styles.request_tr1}
-                            style={{ backgroundColor: "transparent" }}
-                          >
-                            <th className={styles.request_th}>Items</th>
-                            <th
-                              className={
-                                styles.request_th + " " + styles.hideData
-                              }
+                          {matches ? (
+                            <div
+                              className={styles.request_tr1}
+                              style={{ backgroundColor: "transparent" }}
                             >
-                              Quantity
-                            </th>
-                            <th className={styles.request_th}>Set Price</th>
-                            <th className={styles.request_th}>
-                              Product Available
-                            </th>
-                          </div>
+                              <th className={styles.request_th}>Items</th>
+                              <th
+                                className={
+                                  styles.request_th + " " + styles.hideData
+                                }
+                              >
+                                Quantity
+                              </th>
+                              <th className={styles.request_th}>Set Price</th>
+                              <th className={styles.request_th}>
+                                Product Available
+                              </th>
+                            </div>
+                          ) : (
+                            <div
+                              className={styles.request_tr1}
+                              style={{ backgroundColor: "transparent" }}
+                            >
+                              <th className={styles.request_th}>Items</th>
+                              <th
+                                className={
+                                  styles.request_th + " " + styles.hideData
+                                }
+                              >
+                                Quantity
+                              </th>
+                              <th className={styles.request_th}>Set Price</th>
+                            </div>
+                          )}
                         </thead>
                         <tbody className={styles.tbody}>
                           {formState.ingredientsAvailable?.map(
