@@ -6,14 +6,15 @@ import axios from "../../util/Api";
 import { useEffect } from "react";
 import { useMediaQuery } from "../../hooks/usemediaquery";
 import { IoIosSearch } from "react-icons/io";
+import { GoSearch } from "react-icons/go";
 
 export const MobileSearch = ({ setShowDropdown }) => {
   const router = useRouter();
   const [show, setShow] = useState(false);
   const [value, setValue] = useState("");
+  const [chef, setChef] = useState("");
   const [items, setItems] = useState([]);
   const [store, setStore] = useState([]);
-  const matches = useMediaQuery("(min-width: 767px)");
   const [showCategory, setShowCategory] = useState(false);
   const [categories, setCategories] = useState([
     {
@@ -50,6 +51,7 @@ export const MobileSearch = ({ setShowDropdown }) => {
   const [oneStore, setOneStore] = useState({
     visible: false,
     id: "",
+    name: ""
   });
   const getItem = async (name) => {
     try {
@@ -62,6 +64,7 @@ export const MobileSearch = ({ setShowDropdown }) => {
           price: element?.item_price ? `$${element?.item_price}` : "No Price",
           store: element?.store_available?.store_name || "No store",
           item_type: element?.item_type,
+          meal_chef: element?.meal_chef
         };
       });
       setItems(resp);
@@ -71,7 +74,6 @@ export const MobileSearch = ({ setShowDropdown }) => {
     }
     return name;
   };
-  console.log(items, "item");
 
   const getStore = async (name) => {
     try {
@@ -109,27 +111,37 @@ export const MobileSearch = ({ setShowDropdown }) => {
       true
     );
   }, []);
+
   return (
     <div className={styles.two3} ref={ref}>
       <div>
         <div
-          className={styles.searchbox2}
-          onClick={() => setShowCategory(!showCategory)}
+          style={{
+            display: "flex",
+            alignItems: "center",
+            background: "#FFF",
+            width: "100%",
+          }}
         >
-          {categories[0].value && categories.some((ele) => !ele.value)
-            ? "All"
-            : categories.find((ele) => ele.value).label}
-          <GoTriangleUp
-            className={!showCategory ? styles.rotate : styles.nonrotate}
-            size={15}
-          />
+          <div
+            className={styles.cats}
+            onClick={() => setShowCategory(!showCategory)}
+          >
+            {categories[0].value && categories.some((ele) => !ele.value)
+              ? "All"
+              : categories.find((ele) => ele.value).label}
+            <GoTriangleUp
+              className={!showCategory ? styles.rotate : styles.nonrotate}
+              size={15}
+            />
+          </div>
         </div>
 
         {showCategory && (
           <div className={styles.categories}>
             {categories.map((option) => (
               <p
-                style={{ color: "black" }}
+                style={{ color: "rgba(109, 109, 109, 1)" }}
                 onClick={() => {
                   let arr = [];
                   if (option.label === "All") {
@@ -166,9 +178,10 @@ export const MobileSearch = ({ setShowDropdown }) => {
         )}
       </div>
       <div className={styles.searchflex}>
-        <div className={styles.searchfield}>
+        <div className={styles.searchboxfield}>
+          <GoSearch size={12} color="rgba(148, 148, 148, 1)" />
           <input
-            placeholder="Search"
+            placeholder="Search Marketplace"
             autoComplete="off"
             onFocus={() => setShow(true)}
             value={value}
@@ -177,7 +190,7 @@ export const MobileSearch = ({ setShowDropdown }) => {
               getItem(e.target.value);
               getStore(e.target.value);
             }}
-            type="text"
+            type="search"
             name="search"
           />
           {show && (
@@ -212,6 +225,7 @@ export const MobileSearch = ({ setShowDropdown }) => {
                                 setOneStore({
                                   visible: true,
                                   id: stores.value,
+                                  name: stores.label
                                 });
                                 setValue(stores.label);
                               }}
@@ -255,6 +269,7 @@ export const MobileSearch = ({ setShowDropdown }) => {
                                   id: "",
                                 });
                                 setValue(elem.label);
+                                setChef(elem.meal_chef)
                               }}
                               style={{ cursor: "pointer" }}
                             >
@@ -279,7 +294,7 @@ export const MobileSearch = ({ setShowDropdown }) => {
                         Boolean(value) ? (
                           <div className={styles.result}>
                             <p>No Result Found</p>
-                            <button onClick={() => router.push("/suggestmeal")}>
+                            <button onClick={() => router.push("/suggestproduct")}>
                               Suggest Product
                             </button>
                           </div>
@@ -311,43 +326,43 @@ export const MobileSearch = ({ setShowDropdown }) => {
               <>
                 {categories.find((ele) => ele.label === "Kitchen Utensils")
                   ?.value && (
-                  <>
-                    <h4 className={styles.storeTitle}>
-                      Kitchen Utensils ({filteredUtensils().length})
-                    </h4>
-                    <div className={styles.bord} />
-                    <div className={styles.list}>
-                      {filteredUtensils().length === 0 ? (
-                        Boolean(value) ? (
-                          <div className={styles.result}>
-                            <p>No Result Found</p>
-                            <button onClick={() => router.push("/suggestmeal")}>
-                              Suggest Utensil
-                            </button>
-                          </div>
-                        ) : null
-                      ) : (
-                        filteredUtensils()
-                          ?.slice(0, 4)
-                          .map((elem) => (
-                            <p
-                              key={elem.value}
-                              onClick={() => {
-                                setOneStore({
-                                  visible: false,
-                                  id: "",
-                                });
-                                setValue(elem.label);
-                              }}
-                              style={{ cursor: "pointer" }}
-                            >
-                              {elem.label}
-                            </p>
-                          ))
-                      )}
-                    </div>
-                  </>
-                )}
+                    <>
+                      <h4 className={styles.storeTitle}>
+                        Kitchen Utensils ({filteredUtensils().length})
+                      </h4>
+                      <div className={styles.bord} />
+                      <div className={styles.list}>
+                        {filteredUtensils().length === 0 ? (
+                          Boolean(value) ? (
+                            <div className={styles.result}>
+                              <p>No Result Found</p>
+                              <button onClick={() => router.push("/suggestutensil")}>
+                                Suggest Utensil
+                              </button>
+                            </div>
+                          ) : null
+                        ) : (
+                          filteredUtensils()
+                            ?.slice(0, 4)
+                            .map((elem) => (
+                              <p
+                                key={elem.value}
+                                onClick={() => {
+                                  setOneStore({
+                                    visible: false,
+                                    id: "",
+                                  });
+                                  setValue(elem.label);
+                                }}
+                                style={{ cursor: "pointer" }}
+                              >
+                                {elem.label}
+                              </p>
+                            ))
+                        )}
+                      </div>
+                    </>
+                  )}
               </>
             </div>
           )}
@@ -356,13 +371,26 @@ export const MobileSearch = ({ setShowDropdown }) => {
           className={styles.searchbtn}
           onClick={() => {
             if (oneStore.visible) {
-              router.push(`/store/${oneStore.id}`);
+              localStorage.setItem("storeId", oneStore.id)
+              router.push(`/store/${oneStore.name}`);
             } else {
-              items.item_type === "Meal"
-                ? router.push(`/meal/${value}`)
-                : items.item_type === "Product"
-                ? router.push(`/product/${value}`)
-                : router.push(`/product/${value}`);
+              if (
+                items.find(
+                  (ele) =>
+                    ele.item_type === "Meal" && ele.item_type !== "Product"
+                )
+              ) {
+                router.push(`/meal/${chef}/${value}`);
+              } else if (
+                items.filter(
+                  (ele) =>
+                    ele.item_type === "Product" && ele.item_type !== "Meal"
+                )
+              ) {
+                router.push(`/product/${value}`);
+              } else {
+                // router.push(`/meal/${value}`);
+              }
             }
           }}
         >
