@@ -49,6 +49,7 @@ export const SearchDropdown = ({ setShowDropdown }) => {
   const [oneStore, setOneStore] = useState({
     visible: false,
     id: "",
+    name: ""
   });
   const getItem = async (name) => {
     try {
@@ -211,6 +212,7 @@ export const SearchDropdown = ({ setShowDropdown }) => {
                                 setOneStore({
                                   visible: true,
                                   id: stores.value,
+                                  name: stores.label
                                 });
                                 setValue(stores.label);
                               }}
@@ -251,9 +253,10 @@ export const SearchDropdown = ({ setShowDropdown }) => {
                               onClick={() => {
                                 setOneStore({
                                   visible: false,
-                                  id: "",
+                                  id: elem.value,
                                 });
                                 setValue(elem.label);
+                                console.log(oneStore, 'one sttt')
                               }}
                               style={{ cursor: "pointer" }}
                             >
@@ -292,7 +295,7 @@ export const SearchDropdown = ({ setShowDropdown }) => {
                               onClick={() => {
                                 setOneStore({
                                   visible: false,
-                                  id: "",
+                                  id: elem.value,
                                 });
                                 setValue(elem.label);
                               }}
@@ -334,7 +337,7 @@ export const SearchDropdown = ({ setShowDropdown }) => {
                               onClick={() => {
                                 setOneStore({
                                   visible: false,
-                                  id: "",
+                                  id: elem.value,
                                 });
                                 setValue(elem.label);
                               }}
@@ -355,7 +358,8 @@ export const SearchDropdown = ({ setShowDropdown }) => {
           className={styles.searchbtn}
           onClick={() => {
             if (oneStore.visible) {
-              router.push(`/store/${oneStore.id}`);
+              localStorage.setItem("storeId", id)
+              router.push(`/store/${oneStore.name}`)
             } else {
               items.item_type === "Meal"
                 ? router.push(`/meal/${value}`)
@@ -371,3 +375,5 @@ export const SearchDropdown = ({ setShowDropdown }) => {
     </div>
   );
 };
+
+
