@@ -148,7 +148,22 @@ export function StoreMgtIcon(props) {
     </svg>
   );
 }
-
+export function StoreMgtIcon2(props) {
+  return (
+    <svg
+      width="12"
+      height="14"
+      viewBox="0 0 12 14"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M11.3333 1.99935H0.666667V0.666016H11.3333V1.99935ZM12 5.99935V7.33268H11.3333V11.3327H10V7.33268H7.33333V11.3327H0.666667V7.33268H0V5.99935L0.666667 2.66602H11.3333L12 5.99935ZM6 7.33268H2V9.99935H6V7.33268ZM2.66667 13.9993H4V12.666H2.66667V13.9993ZM5.33333 13.9993H6.66667V12.666H5.33333V13.9993ZM8 13.9993H9.33333V12.666H8V13.9993Z"
+        fill="white"
+      />
+    </svg>
+  );
+}
 export function SupportIcon(props) {
   return (
     <svg
@@ -183,6 +198,38 @@ export function CartIcon(props) {
       onClick={props.cartOpen}
     >
       <path d="M11.994 4.03538L11.4907 6.92234C11.4287 7.24125 11.1746 7.43516 10.8816 7.4388H3.545L3.39934 8.25986H10.3916C10.7598 8.28784 11.0095 8.54866 11.014 8.88226C10.9872 9.24969 10.7264 9.50012 10.3916 9.50468H2.65773C2.23946 9.46706 1.99005 9.13365 2.03531 8.76308L2.36638 6.96204L1.86315 1.90324L0.432919 1.45301C0.083409 1.31424 -0.0592411 0.998062 0.0223889 0.671682C0.158119 0.331532 0.483349 0.176122 0.803719 0.261152L2.63124 0.843842C2.87744 0.936132 3.01907 1.13766 3.05501 1.37356L3.16095 2.38002L11.451 3.30702C11.8319 3.38953 12.0387 3.69124 11.994 4.03538ZM4.55272 10.8116C4.55272 11.3363 4.12732 11.7617 3.60254 11.7617C3.07777 11.7617 2.65237 11.3363 2.65237 10.8116C2.65237 10.2868 3.07778 9.86138 3.60254 9.86138C4.12731 9.86139 4.55272 10.2868 4.55272 10.8116ZM10.2252 10.8116C10.2252 11.3363 9.79983 11.7617 9.27505 11.7617C8.75028 11.7617 8.32488 11.3363 8.32488 10.8116C8.32488 10.2868 8.75028 9.86138 9.27505 9.86138C9.79983 9.86139 10.2252 10.2868 10.2252 10.8116Z" />
+    </svg>
+  );
+}
+export function CartIcon2({ color }) {
+  return (
+    <svg
+      width="12"
+      height="14"
+      viewBox="0 0 12 12"
+      xmlns="http://www.w3.org/2000/svg"
+    // onClick={props.cartOpen}
+    >
+      <path
+        fill={color}
+        d="M11.994 4.03538L11.4907 6.92234C11.4287 7.24125 11.1746 7.43516 10.8816 7.4388H3.545L3.39934 8.25986H10.3916C10.7598 8.28784 11.0095 8.54866 11.014 8.88226C10.9872 9.24969 10.7264 9.50012 10.3916 9.50468H2.65773C2.23946 9.46706 1.99005 9.13365 2.03531 8.76308L2.36638 6.96204L1.86315 1.90324L0.432919 1.45301C0.083409 1.31424 -0.0592411 0.998062 0.0223889 0.671682C0.158119 0.331532 0.483349 0.176122 0.803719 0.261152L2.63124 0.843842C2.87744 0.936132 3.01907 1.13766 3.05501 1.37356L3.16095 2.38002L11.451 3.30702C11.8319 3.38953 12.0387 3.69124 11.994 4.03538ZM4.55272 10.8116C4.55272 11.3363 4.12732 11.7617 3.60254 11.7617C3.07777 11.7617 2.65237 11.3363 2.65237 10.8116C2.65237 10.2868 3.07778 9.86138 3.60254 9.86138C4.12731 9.86139 4.55272 10.2868 4.55272 10.8116ZM10.2252 10.8116C10.2252 11.3363 9.79983 11.7617 9.27505 11.7617C8.75028 11.7617 8.32488 11.3363 8.32488 10.8116C8.32488 10.2868 8.75028 9.86138 9.27505 9.86138C9.79983 9.86139 10.2252 10.2868 10.2252 10.8116Z"
+      />
+    </svg>
+  );
+}
+export function FaqIcon(props) {
+  return (
+    <svg
+      width="18"
+      height="18"
+      viewBox="0 0 21 21"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M10.5 19.25L10.2813 16.625H10.0625C7.99167 16.625 6.23438 15.9031 4.79063 14.4594C3.34688 13.0156 2.625 11.2583 2.625 9.1875C2.625 7.11667 3.34688 5.35938 4.79063 3.91563C6.23438 2.47188 7.99167 1.75 10.0625 1.75C11.0979 1.75 12.0639 1.94308 12.9605 2.32925C13.8577 2.716 14.6455 3.24829 15.3239 3.92613C16.0017 4.60454 16.5337 5.39204 16.9199 6.28863C17.3066 7.18579 17.5 8.15208 17.5 9.1875C17.5 10.2813 17.3215 11.3313 16.9645 12.3375C16.6069 13.3438 16.1184 14.2771 15.4989 15.1375C14.8788 15.9979 14.1423 16.7781 13.2895 17.4781C12.4361 18.1781 11.5063 18.7688 10.5 19.25ZM10.0406 13.9781C10.2885 13.9781 10.5 13.8906 10.675 13.7156C10.85 13.5406 10.9375 13.3292 10.9375 13.0813C10.9375 12.8333 10.85 12.6219 10.675 12.4469C10.5 12.2719 10.2885 12.1844 10.0406 12.1844C9.79271 12.1844 9.58125 12.2719 9.40625 12.4469C9.23125 12.6219 9.14375 12.8333 9.14375 13.0813C9.14375 13.3292 9.23125 13.5406 9.40625 13.7156C9.58125 13.8906 9.79271 13.9781 10.0406 13.9781ZM9.40625 11.2H10.7188C10.7188 10.7625 10.7625 10.4563 10.85 10.2813C10.9375 10.1063 11.2146 9.78542 11.6813 9.31875C11.9438 9.05625 12.1625 8.77188 12.3375 8.46563C12.5125 8.15938 12.6 7.83125 12.6 7.48125C12.6 6.7375 12.3486 6.17954 11.8458 5.80738C11.3423 5.43579 10.7479 5.25 10.0625 5.25C9.42083 5.25 8.88125 5.4285 8.44375 5.7855C8.00625 6.14308 7.7 6.57708 7.525 7.0875L8.75 7.56875C8.82292 7.32083 8.96146 7.07642 9.16563 6.8355C9.36979 6.59517 9.66875 6.475 10.0625 6.475C10.4563 6.475 10.7517 6.58438 10.9489 6.80313C11.1455 7.02188 11.2438 7.2625 11.2438 7.525C11.2438 7.77292 11.1708 7.99517 11.025 8.19175C10.8792 8.38892 10.7042 8.58958 10.5 8.79375C9.98958 9.23125 9.67983 9.57746 9.57075 9.83238C9.46108 10.0879 9.40625 10.5438 9.40625 11.2Z"
+        fill="#6D6D6D"
+      />
     </svg>
   );
 }
@@ -247,6 +294,22 @@ export function HomeIcon(props) {
     </svg>
   );
 }
+export function HomeIcon2({ color }) {
+  return (
+    <svg
+      width="14"
+      height="13"
+      viewBox="0 0 14 13"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M0.333984 6.79092L6.52932 0.595582C6.65434 0.470601 6.82387 0.400391 7.00065 0.400391C7.17743 0.400391 7.34697 0.470601 7.47198 0.595582L13.6673 6.79092H12.334V12.1242C12.334 12.3011 12.2637 12.4706 12.1387 12.5957C12.0137 12.7207 11.8441 12.7909 11.6673 12.7909H8.33398V8.12425H5.66732V12.7909H2.33398C2.15717 12.7909 1.9876 12.7207 1.86258 12.5957C1.73756 12.4706 1.66732 12.3011 1.66732 12.1242V6.79092H0.333984Z"
+        fill={color}
+      />
+    </svg>
+  );
+}
 
 export function Order2Icon(props) {
   return (
@@ -259,6 +322,21 @@ export function Order2Icon(props) {
     </svg>
   );
 }
+export function Order3Icon({ color }) {
+  return (
+    <svg
+      height="13"
+      width="14"
+      viewBox="0 0 15 13"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        fill={color}
+        d="M8.53946 10.0024L7.66046 7.93741L6.65696 10.3139L2.89696 8.66891L2.98446 10.6219L7.70296 12.8089L13.158 10.1264L13.083 7.63141L8.53946 10.0024ZM14.7275 2.31041L9.26046 0.691406L7.75796 2.02641L6.21296 0.847406L0.772461 3.10041L2.67046 4.62391L0.955461 7.27541L6.39546 9.65341L7.66046 6.65441L8.78846 9.30841L14.485 6.33741L12.9325 3.98241L14.7285 2.30991L14.7275 2.31041ZM7.63896 6.29041L3.10896 4.37541L7.74696 2.32441L12.5505 3.95191L7.63896 6.28991V6.29041Z"
+      />
+    </svg>
+  );
+}
 
 export function BasketIcon(props) {
   return (
@@ -268,6 +346,22 @@ export function BasketIcon(props) {
       xmlns="http://www.w3.org/2000/svg"
     >
       <path d="M14.0625 4.71429C14.3213 4.71429 14.5422 4.79102 14.7253 4.94448C14.9084 5.09794 15 5.28311 15 5.5C15 5.71689 14.9084 5.90206 14.7253 6.05552C14.5422 6.20898 14.3213 6.28571 14.0625 6.28571H13.9526L13.1104 10.3493C13.0713 10.5376 12.9639 10.6931 12.7881 10.8158C12.6123 10.9386 12.4121 11 12.1875 11H2.8125C2.58789 11 2.3877 10.9386 2.21191 10.8158C2.03613 10.6931 1.92871 10.5376 1.88965 10.3493L1.04736 6.28571H0.9375C0.678711 6.28571 0.457764 6.20898 0.274658 6.05552C0.0915527 5.90206 0 5.71689 0 5.5C0 5.28311 0.0915527 5.09794 0.274658 4.94448C0.457764 4.79102 0.678711 4.71429 0.9375 4.71429H14.0625ZM3.55225 9.625C3.6792 9.61682 3.7854 9.57078 3.87085 9.48689C3.9563 9.40299 3.99414 9.30785 3.98438 9.20145L3.75 6.64788C3.74023 6.54148 3.6853 6.45247 3.58521 6.38086C3.48511 6.30924 3.37158 6.27753 3.24463 6.28571C3.11768 6.2939 3.01147 6.33994 2.92603 6.42383C2.84058 6.50772 2.80273 6.60286 2.8125 6.70926L3.04688 9.26283C3.05664 9.36514 3.10669 9.45108 3.19702 9.52065C3.28735 9.59022 3.39355 9.625 3.51562 9.625H3.55225ZM6.5625 9.23214V6.67857C6.5625 6.57217 6.51611 6.4801 6.42334 6.40234C6.33057 6.32459 6.2207 6.28571 6.09375 6.28571C5.9668 6.28571 5.85693 6.32459 5.76416 6.40234C5.67139 6.4801 5.625 6.57217 5.625 6.67857V9.23214C5.625 9.33854 5.67139 9.43062 5.76416 9.50837C5.85693 9.58612 5.9668 9.625 6.09375 9.625C6.2207 9.625 6.33057 9.58612 6.42334 9.50837C6.51611 9.43062 6.5625 9.33854 6.5625 9.23214ZM9.375 9.23214V6.67857C9.375 6.57217 9.32861 6.4801 9.23584 6.40234C9.14307 6.32459 9.0332 6.28571 8.90625 6.28571C8.7793 6.28571 8.66943 6.32459 8.57666 6.40234C8.48389 6.4801 8.4375 6.57217 8.4375 6.67857V9.23214C8.4375 9.33854 8.48389 9.43062 8.57666 9.50837C8.66943 9.58612 8.7793 9.625 8.90625 9.625C9.0332 9.625 9.14307 9.58612 9.23584 9.50837C9.32861 9.43062 9.375 9.33854 9.375 9.23214ZM11.9531 9.26283L12.1875 6.70926C12.1973 6.60286 12.1594 6.50772 12.074 6.42383C11.9885 6.33994 11.8823 6.2939 11.7554 6.28571C11.6284 6.27753 11.5149 6.30924 11.4148 6.38086C11.3147 6.45247 11.2598 6.54148 11.25 6.64788L11.0156 9.20145C11.0059 9.30785 11.0437 9.40299 11.1292 9.48689C11.2146 9.57078 11.3208 9.61682 11.4478 9.625H11.4844C11.6064 9.625 11.7126 9.59022 11.803 9.52065C11.8933 9.45108 11.9434 9.36514 11.9531 9.26283ZM3.48633 1.79241L2.80518 4.32143H1.83838L2.57812 1.6144C2.6709 1.25428 2.88818 0.960658 3.22998 0.733538C3.57178 0.506417 3.9624 0.392857 4.40186 0.392857H5.625C5.625 0.286458 5.67139 0.194382 5.76416 0.116629C5.85693 0.0388765 5.9668 0 6.09375 0H8.90625C9.0332 0 9.14307 0.0388765 9.23584 0.116629C9.32861 0.194382 9.375 0.286458 9.375 0.392857H10.5981C11.0376 0.392857 11.4282 0.506417 11.77 0.733538C12.1118 0.960658 12.3291 1.25428 12.4219 1.6144L13.1616 4.32143H12.1948L11.5137 1.79241C11.46 1.61235 11.3489 1.46503 11.1804 1.35045C11.012 1.23586 10.8179 1.17857 10.5981 1.17857H9.375C9.375 1.28497 9.32861 1.37705 9.23584 1.4548C9.14307 1.53255 9.0332 1.57143 8.90625 1.57143H6.09375C5.9668 1.57143 5.85693 1.53255 5.76416 1.4548C5.67139 1.37705 5.625 1.28497 5.625 1.17857H4.40186C4.18213 1.17857 3.98804 1.23586 3.81958 1.35045C3.65112 1.46503 3.54004 1.61235 3.48633 1.79241Z" />
+    </svg>
+  );
+}
+export function BasketIcon2({ color }) {
+  return (
+    <svg
+      width="15"
+      height="11"
+      viewBox="0 0 15 11"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M14.0625 4.71429C14.3213 4.71429 14.5422 4.79102 14.7253 4.94448C14.9084 5.09794 15 5.28311 15 5.5C15 5.71689 14.9084 5.90206 14.7253 6.05552C14.5422 6.20898 14.3213 6.28571 14.0625 6.28571H13.9526L13.1104 10.3493C13.0713 10.5376 12.9639 10.6931 12.7881 10.8158C12.6123 10.9386 12.4121 11 12.1875 11H2.8125C2.58789 11 2.3877 10.9386 2.21191 10.8158C2.03613 10.6931 1.92871 10.5376 1.88965 10.3493L1.04736 6.28571H0.9375C0.678711 6.28571 0.457764 6.20898 0.274658 6.05552C0.0915527 5.90206 0 5.71689 0 5.5C0 5.28311 0.0915527 5.09794 0.274658 4.94448C0.457764 4.79102 0.678711 4.71429 0.9375 4.71429H14.0625ZM3.55225 9.625C3.6792 9.61682 3.7854 9.57078 3.87085 9.48689C3.9563 9.40299 3.99414 9.30785 3.98438 9.20145L3.75 6.64788C3.74023 6.54148 3.6853 6.45247 3.58521 6.38086C3.48511 6.30924 3.37158 6.27753 3.24463 6.28571C3.11768 6.2939 3.01147 6.33994 2.92603 6.42383C2.84058 6.50772 2.80273 6.60286 2.8125 6.70926L3.04688 9.26283C3.05664 9.36514 3.10669 9.45108 3.19702 9.52065C3.28735 9.59022 3.39355 9.625 3.51562 9.625H3.55225ZM6.5625 9.23214V6.67857C6.5625 6.57217 6.51611 6.4801 6.42334 6.40234C6.33057 6.32459 6.2207 6.28571 6.09375 6.28571C5.9668 6.28571 5.85693 6.32459 5.76416 6.40234C5.67139 6.4801 5.625 6.57217 5.625 6.67857V9.23214C5.625 9.33854 5.67139 9.43062 5.76416 9.50837C5.85693 9.58612 5.9668 9.625 6.09375 9.625C6.2207 9.625 6.33057 9.58612 6.42334 9.50837C6.51611 9.43062 6.5625 9.33854 6.5625 9.23214ZM9.375 9.23214V6.67857C9.375 6.57217 9.32861 6.4801 9.23584 6.40234C9.14307 6.32459 9.0332 6.28571 8.90625 6.28571C8.7793 6.28571 8.66943 6.32459 8.57666 6.40234C8.48389 6.4801 8.4375 6.57217 8.4375 6.67857V9.23214C8.4375 9.33854 8.48389 9.43062 8.57666 9.50837C8.66943 9.58612 8.7793 9.625 8.90625 9.625C9.0332 9.625 9.14307 9.58612 9.23584 9.50837C9.32861 9.43062 9.375 9.33854 9.375 9.23214ZM11.9531 9.26283L12.1875 6.70926C12.1973 6.60286 12.1594 6.50772 12.074 6.42383C11.9885 6.33994 11.8823 6.2939 11.7554 6.28571C11.6284 6.27753 11.5149 6.30924 11.4148 6.38086C11.3147 6.45247 11.2598 6.54148 11.25 6.64788L11.0156 9.20145C11.0059 9.30785 11.0437 9.40299 11.1292 9.48689C11.2146 9.57078 11.3208 9.61682 11.4478 9.625H11.4844C11.6064 9.625 11.7126 9.59022 11.803 9.52065C11.8933 9.45108 11.9434 9.36514 11.9531 9.26283ZM3.48633 1.79241L2.80518 4.32143H1.83838L2.57812 1.6144C2.6709 1.25428 2.88818 0.960658 3.22998 0.733538C3.57178 0.506417 3.9624 0.392857 4.40186 0.392857H5.625C5.625 0.286458 5.67139 0.194382 5.76416 0.116629C5.85693 0.0388765 5.9668 0 6.09375 0H8.90625C9.0332 0 9.14307 0.0388765 9.23584 0.116629C9.32861 0.194382 9.375 0.286458 9.375 0.392857H10.5981C11.0376 0.392857 11.4282 0.506417 11.77 0.733538C12.1118 0.960658 12.3291 1.25428 12.4219 1.6144L13.1616 4.32143H12.1948L11.5137 1.79241C11.46 1.61235 11.3489 1.46503 11.1804 1.35045C11.012 1.23586 10.8179 1.17857 10.5981 1.17857H9.375C9.375 1.28497 9.32861 1.37705 9.23584 1.4548C9.14307 1.53255 9.0332 1.57143 8.90625 1.57143H6.09375C5.9668 1.57143 5.85693 1.53255 5.76416 1.4548C5.67139 1.37705 5.625 1.28497 5.625 1.17857H4.40186C4.18213 1.17857 3.98804 1.23586 3.81958 1.35045C3.65112 1.46503 3.54004 1.61235 3.48633 1.79241Z"
+        fill={color}
+      />
     </svg>
   );
 }
@@ -420,7 +514,7 @@ export function LocationIcon(props) {
 
 export function ShareIcon(props) {
   return (
-    <svg viewBox="0 0 13 13" xmlns="http://www.w3.org/2000/svg">
+    <svg className={props.style} viewBox="0 0 13 13" xmlns="http://www.w3.org/2000/svg">
       <path d="M9.95974 8.63502C9.69781 8.63602 9.43879 8.68993 9.19821 8.79351C8.95764 8.89709 8.74047 9.04821 8.55974 9.2378L3.81918 6.86558C3.9358 6.51198 3.9358 6.13029 3.81918 5.77669L8.56752 3.36169C8.90763 3.7201 9.36886 3.93925 9.86155 3.97654C10.3542 4.01384 10.8432 3.86661 11.2334 3.56348C11.6236 3.26035 11.8871 2.82297 11.9728 2.33635C12.0585 1.84974 11.9601 1.34865 11.6969 0.930486C11.4337 0.512321 11.0245 0.206954 10.5487 0.0737368C10.0729 -0.0594801 9.56453 -0.0110291 9.12246 0.209672C8.68039 0.430374 8.33617 0.807559 8.1567 1.26791C7.97724 1.72827 7.97535 2.23891 8.15141 2.70058L3.44974 5.09225C3.19581 4.7817 2.85205 4.55729 2.46558 4.44977C2.07911 4.34226 1.66885 4.35691 1.29103 4.49171C0.913211 4.62651 0.586332 4.87487 0.355201 5.20274C0.124069 5.5306 0 5.92193 0 6.32308C0 6.72423 0.124069 7.11555 0.355201 7.44342C0.586332 7.77129 0.913211 8.01965 1.29103 8.15445C1.66885 8.28925 2.07911 8.3039 2.46558 8.19639C2.85205 8.08887 3.19581 7.86446 3.44974 7.55391L8.13974 9.91447C8.06049 10.1272 8.01968 10.3524 8.01919 10.5795C8.01919 10.964 8.13322 11.34 8.34688 11.6597C8.56054 11.9795 8.86422 12.2287 9.21952 12.3759C9.57482 12.5231 9.96579 12.5616 10.343 12.4866C10.7202 12.4115 11.0666 12.2263 11.3386 11.9544C11.6105 11.6825 11.7957 11.336 11.8707 10.9588C11.9457 10.5816 11.9072 10.1907 11.7601 9.83536C11.6129 9.48006 11.3637 9.17638 11.0439 8.96272C10.7241 8.74906 10.3482 8.63502 9.96363 8.63502H9.95974Z" />
     </svg>
   );
@@ -632,12 +726,169 @@ export function ArrowUp2Icon(props) {
 
 export function ChatIcon(props) {
   return (
-    <svg viewBox="0 0 20 17" xmlns="http://www.w3.org/2000/svg">
-      <path d="M20 7.08333C20 8.36719 19.5536 9.55328 18.6607 10.6416C17.7679 11.7299 16.5551 12.5895 15.0223 13.2204C13.4896 13.8512 11.8155 14.1667 10 14.1667C9.47917 14.1667 8.93973 14.1372 8.3817 14.0781C6.90848 15.3694 5.19717 16.2622 3.24777 16.7565C2.88318 16.8598 2.45908 16.941 1.97545 17C1.84896 17.0148 1.73549 16.9816 1.63504 16.9004C1.5346 16.8192 1.46949 16.7122 1.43973 16.5794V16.5684C1.41741 16.5388 1.41555 16.4946 1.43415 16.4355C1.45275 16.3765 1.46019 16.3396 1.45647 16.3249C1.45275 16.3101 1.46949 16.2751 1.5067 16.2197L1.57366 16.1201L1.65179 16.026L1.74107 15.9264C1.79315 15.8674 1.90848 15.7401 2.08705 15.5446C2.26562 15.3491 2.39397 15.2089 2.4721 15.124C2.55022 15.0392 2.66555 14.8934 2.81808 14.6868C2.97061 14.4803 3.09152 14.2921 3.1808 14.1224C3.27009 13.9527 3.37054 13.735 3.48214 13.4694C3.59375 13.2038 3.69048 12.9234 3.77232 12.6283C2.60417 11.9716 1.68341 11.1599 1.01004 10.1934C0.336682 9.22678 0 8.1901 0 7.08333C0 6.12413 0.264137 5.20736 0.792411 4.33301C1.32068 3.45866 2.03125 2.70421 2.92411 2.06966C3.81696 1.43511 4.88095 0.931532 6.11607 0.558919C7.35119 0.186306 8.64583 0 10 0C11.8155 0 13.4896 0.31543 15.0223 0.946289C16.5551 1.57715 17.7679 2.43674 18.6607 3.52507C19.5536 4.61339 20 5.79948 20 7.08333Z" />
+    <svg
+      className={props.style}
+      width="12"
+      height="12"
+      viewBox="0 0 12 12"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clip-path="url(#clip0_7833_1233)">
+        <path
+          d="M12 6C12 8.8995 9.3135 11.25 6 11.25C5.40574 11.2508 4.81394 11.1736 4.23975 11.0205C3.80175 11.2425 2.796 11.6685 1.104 11.946C0.954 11.97 0.84 11.814 0.89925 11.6745C1.16475 11.0475 1.40475 10.212 1.47675 9.45C0.558 8.5275 0 7.32 0 6C0 3.1005 2.6865 0.75 6 0.75C9.3135 0.75 12 3.1005 12 6ZM3.375 3.75C3.27554 3.75 3.18016 3.78951 3.10984 3.85984C3.03951 3.93016 3 4.02554 3 4.125C3 4.22446 3.03951 4.31984 3.10984 4.39016C3.18016 4.46049 3.27554 4.5 3.375 4.5H8.625C8.72446 4.5 8.81984 4.46049 8.89017 4.39016C8.96049 4.31984 9 4.22446 9 4.125C9 4.02554 8.96049 3.93016 8.89017 3.85984C8.81984 3.78951 8.72446 3.75 8.625 3.75H3.375ZM3.375 5.625C3.27554 5.625 3.18016 5.66451 3.10984 5.73484C3.03951 5.80516 3 5.90054 3 6C3 6.09946 3.03951 6.19484 3.10984 6.26517C3.18016 6.33549 3.27554 6.375 3.375 6.375H8.625C8.72446 6.375 8.81984 6.33549 8.89017 6.26517C8.96049 6.19484 9 6.09946 9 6C9 5.90054 8.96049 5.80516 8.89017 5.73484C8.81984 5.66451 8.72446 5.625 8.625 5.625H3.375ZM3.375 7.5C3.27554 7.5 3.18016 7.53951 3.10984 7.60983C3.03951 7.68016 3 7.77554 3 7.875C3 7.97446 3.03951 8.06984 3.10984 8.14017C3.18016 8.21049 3.27554 8.25 3.375 8.25H6.375C6.47446 8.25 6.56984 8.21049 6.64017 8.14017C6.71049 8.06984 6.75 7.97446 6.75 7.875C6.75 7.77554 6.71049 7.68016 6.64017 7.60983C6.56984 7.53951 6.47446 7.5 6.375 7.5H3.375Z"
+          fill="white"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_7833_1233">
+          <rect width="12" height="12" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
+export function AdminMgtIcon(props) {
+  return (
+    <svg
+      width="17"
+      height="17"
+      viewBox="0 0 15 15"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M7.5 1.875C8.88125 1.875 10 2.99375 10 4.375C10 5.75625 8.88125 6.875 7.5 6.875C6.11875 6.875 5 5.75625 5 4.375C5 2.99375 6.11875 1.875 7.5 1.875ZM10 8.4625C10 9.125 9.825 10.6687 8.63125 12.3938L8.125 9.375L8.7125 8.2C8.325 8.15625 7.91875 8.125 7.5 8.125C7.08125 8.125 6.675 8.15625 6.2875 8.2L6.875 9.375L6.36875 12.3938C5.175 10.6687 5 9.125 5 8.4625C3.50625 8.9 2.5 9.6875 2.5 10.625V13.125H12.5V10.625C12.5 9.6875 11.5 8.9 10 8.4625Z"
+        fill="white"
+      />
     </svg>
   );
 }
 
+export function BlogSettingsIcon() {
+  return (
+    <svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M0 0V12H12V0H0ZM10 10H2V9.33333H10V10ZM10 8.66667H2V8H10V8.66667ZM10 6H2V2H10V6Z" fill="white" />
+    </svg>
+  )
+}
+
+export function FBIcon() {
+  return (
+    <svg width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <circle cx="15" cy="15" r="14.5" stroke="#1877F2" />
+      <g clip-path="url(#clip0_33_3081)">
+        <path d="M23 15C23 10.5817 19.4183 7 15 7C10.5817 7 7 10.5817 7 15C7 18.993 9.9255 22.3027 13.75 22.9028V17.3125H11.7188V15H13.75V13.2375C13.75 11.2325 14.9444 10.125 16.7717 10.125C17.647 10.125 18.5625 10.2812 18.5625 10.2812V12.25H17.5538C16.5599 12.25 16.25 12.8667 16.25 13.4994V15H18.4688L18.1141 17.3125H16.25V22.9028C20.0745 22.3027 23 18.9931 23 15Z" fill="#1877F2" />
+      </g>
+      <defs>
+        <clipPath id="clip0_33_3081">
+          <rect width="16" height="16" fill="white" transform="translate(7 7)" />
+        </clipPath>
+      </defs>
+    </svg>
+
+  )
+}
+
+export function TwitterIcon() {
+  return (
+    <svg width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <g clip-path="url(#clip0_33_3089)">
+    <path d="M23 10.4612C22.4758 10.7051 21.9197 10.8651 21.3503 10.9361C21.9434 10.5629 22.3988 9.97188 22.6132 9.26773C22.0494 9.619 21.4326 9.86648 20.7893 9.9995C20.2654 9.41345 19.5189 9.0472 18.6928 9.0472C17.1065 9.0472 15.8204 10.3974 15.8204 12.0627C15.8204 12.2991 15.8458 12.5292 15.8948 12.75C13.5077 12.6242 11.3913 11.4237 9.97459 9.5992C9.7274 10.0446 9.58576 10.5626 9.58576 11.1152C9.58576 12.1615 10.0929 13.0844 10.8635 13.6252C10.4074 13.6102 9.96132 13.4809 9.56252 13.248C9.56235 13.2606 9.56235 13.2733 9.56235 13.286C9.56235 14.747 10.5524 15.9659 11.8663 16.2429C11.4434 16.3637 10.9997 16.3813 10.5693 16.2946C10.9347 17.4926 11.9955 18.3644 13.2523 18.3888C12.2693 19.1976 11.0308 19.6797 9.68518 19.6797C9.4533 19.6797 9.22471 19.6654 9 19.6376C10.2711 20.4932 11.7809 20.9924 13.4029 20.9924C18.6861 20.9924 21.5751 16.3973 21.5751 12.4124C21.5751 12.2816 21.5723 12.1516 21.5668 12.0223C22.1291 11.5955 22.6144 11.0669 23 10.4612Z" fill="#55ACEE"/>
+    </g>
+    <circle cx="15" cy="15" r="14.5" stroke="#55ACEE"/>
+    <defs>
+    <clipPath id="clip0_33_3089">
+    <rect width="14" height="12" fill="white" transform="translate(9 9)"/>
+    </clipPath>
+    </defs>
+    </svg>    
+  )
+}
+
+
+export function InstaIcon() {
+  return (
+    <svg width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+<g clip-path="url(#clip0_33_3085)">
+<path d="M14.9993 9C13.3684 9 13.1651 9.00788 12.5269 9.03488C11.886 9.06638 11.4525 9.16575 11.07 9.31425C10.6697 9.46538 10.3071 9.70201 10.0076 10.0076C9.70115 10.3064 9.4644 10.6692 9.31425 11.07C9.16575 11.4525 9.06638 11.886 9.03488 12.5269C9.006 13.1655 9 13.3684 9 14.9992C9 16.6301 9.00788 16.8334 9.03488 17.4716C9.06638 18.1106 9.16575 18.546 9.31425 18.9285C9.46538 19.3288 9.70201 19.6914 10.0076 19.9909C10.3064 20.2973 10.6692 20.5341 11.07 20.6842C11.4525 20.8309 11.8879 20.9321 12.5269 20.9636C13.1655 20.9925 13.3684 20.9985 14.9993 20.9985C16.6301 20.9985 16.8334 20.9906 17.4716 20.9636C18.1106 20.9321 18.546 20.8305 18.9285 20.6842C19.3288 20.5331 19.6914 20.2965 19.9909 19.9909C20.2977 19.6924 20.5346 19.3296 20.6843 18.9285C20.8309 18.546 20.9321 18.1106 20.9636 17.4716C20.9925 16.833 20.9985 16.6301 20.9985 14.9992C20.9985 13.3684 20.9906 13.1651 20.9636 12.5269C20.9321 11.8879 20.8305 11.4502 20.6843 11.07C20.5331 10.6697 20.2965 10.3071 19.9909 10.0076C19.6924 9.70076 19.3296 9.46394 18.9285 9.31425C18.546 9.16575 18.1106 9.06638 17.4716 9.03488C16.833 9.006 16.6301 9 14.9993 9ZM14.9993 10.08C16.6009 10.08 16.7921 10.0879 17.4251 10.1149C18.009 10.1423 18.327 10.2398 18.5381 10.3223C18.7991 10.4186 19.0354 10.5719 19.2296 10.7711C19.4285 10.9646 19.5812 11.2003 19.6766 11.4607C19.7591 11.6719 19.8566 11.9899 19.884 12.5738C19.911 13.2067 19.9189 13.3984 19.9189 14.9996C19.9189 16.6009 19.911 16.7925 19.8818 17.4255C19.8503 18.0094 19.7528 18.3274 19.6706 18.5385C19.557 18.8201 19.4303 19.017 19.221 19.23C19.0261 19.4279 18.79 19.5805 18.5295 19.677C18.321 19.7595 17.9985 19.857 17.4124 19.8844C16.776 19.9114 16.5885 19.9193 14.9828 19.9193C13.377 19.9193 13.1899 19.9114 12.5528 19.8821C11.9689 19.8506 11.6468 19.7531 11.4356 19.671C11.1506 19.5574 10.9553 19.4306 10.746 19.2214C10.5349 19.0103 10.4006 18.8092 10.2971 18.5299C10.2131 18.3214 10.1171 17.9989 10.086 17.4128C10.0646 16.7839 10.0545 16.5889 10.0545 14.9906C10.0545 13.3931 10.0646 13.1978 10.086 12.561C10.1171 11.9749 10.2131 11.6531 10.2971 11.4439C10.4006 11.1585 10.5353 10.9635 10.746 10.7524C10.9549 10.5435 11.1506 10.4085 11.4356 10.3027C11.6468 10.2206 11.961 10.1231 12.5471 10.0939C13.1835 10.071 13.371 10.0628 14.9745 10.0628L14.9993 10.08ZM14.9993 11.9201C14.5948 11.92 14.1942 11.9995 13.8205 12.1543C13.4467 12.309 13.1071 12.5358 12.8211 12.8219C12.5351 13.1079 12.3082 13.4475 12.1535 13.8212C11.9988 14.1949 11.9192 14.5955 11.9194 15C11.9192 15.4045 11.9988 15.8051 12.1535 16.1788C12.3082 16.5525 12.5351 16.8921 12.8211 17.1781C13.1071 17.4642 13.4467 17.691 13.8205 17.8457C14.1942 18.0005 14.5948 18.08 14.9993 18.0799C15.4037 18.08 15.8043 18.0005 16.178 17.8457C16.5518 17.691 16.8914 17.4642 17.1774 17.1781C17.4634 16.8921 17.6903 16.5525 17.845 16.1788C17.9997 15.8051 18.0793 15.4045 18.0791 15C18.0793 14.5955 17.9997 14.1949 17.845 13.8212C17.6903 13.4475 17.4634 13.1079 17.1774 12.8219C16.8914 12.5358 16.5518 12.309 16.178 12.1543C15.8043 11.9995 15.4037 11.92 14.9993 11.9201ZM14.9993 16.9999C13.8938 16.9999 12.9994 16.1055 12.9994 15C12.9994 13.8945 13.8938 13.0001 14.9993 13.0001C16.1048 13.0001 16.9991 13.8945 16.9991 15C16.9991 16.1055 16.1048 16.9999 14.9993 16.9999ZM18.9233 11.7971C18.9231 11.9882 18.847 12.1715 18.7118 12.3066C18.5766 12.4416 18.3932 12.5175 18.2021 12.5175C18.1076 12.5175 18.014 12.4989 17.9267 12.4627C17.8394 12.4266 17.7601 12.3735 17.6933 12.3067C17.6265 12.2399 17.5734 12.1606 17.5373 12.0733C17.5011 11.986 17.4825 11.8924 17.4825 11.7979C17.4825 11.7034 17.5011 11.6098 17.5373 11.5225C17.5734 11.4352 17.6265 11.3558 17.6933 11.289C17.7601 11.2222 17.8394 11.1692 17.9267 11.133C18.014 11.0969 18.1076 11.0782 18.2021 11.0782C18.5985 11.0782 18.9233 11.4007 18.9233 11.7971Z" fill="#F47900"/>
+</g>
+<circle cx="15" cy="15" r="14.5" stroke="#F47900"/>
+<defs>
+<clipPath id="clip0_33_3085">
+<rect width="12" height="12" fill="white" transform="translate(9 9)"/>
+</clipPath>
+</defs>
+</svg>
+   
+  )
+}
+
+
+export function WhappIcon() {
+  return (
+    <svg width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+<circle cx="15" cy="15" r="14.5" stroke="#3ABF4C"/>
+<g clip-path="url(#clip0_33_3093)">
+<path d="M6.38492 14.8922C6.3845 16.4046 6.78275 17.8813 7.54002 19.1828L6.3125 23.63L10.8991 22.4367C12.1677 23.1219 13.5891 23.481 15.0335 23.4811H15.0373C19.8055 23.4811 23.687 19.6311 23.689 14.899C23.69 12.6059 22.7908 10.4497 21.1572 8.82748C19.5238 7.20538 17.3515 6.31159 15.0369 6.31055C10.2681 6.31055 6.38696 10.1603 6.38499 14.8922" fill="url(#paint0_linear_33_3093)"/>
+<path d="M6.07523 14.8895C6.07474 16.4563 6.48727 17.9858 7.27153 19.334L6 23.9405L10.7511 22.7044C12.0602 23.4126 13.5341 23.786 15.0338 23.7866H15.0377C19.977 23.7866 23.9979 19.7981 24 14.8966C24.0008 12.5212 23.0693 10.2874 21.3773 8.60707C19.6851 6.92693 17.4351 6.00098 15.0377 6C10.0975 6 6.07727 9.98791 6.0753 14.8895H6.07523ZM8.90461 19.1018L8.72721 18.8223C7.98148 17.6458 7.58787 16.2862 7.58843 14.89C7.59005 10.816 10.9316 7.5014 15.0405 7.5014C17.0303 7.50223 18.9004 8.27191 20.3069 9.66837C21.7134 11.065 22.4873 12.9215 22.4868 14.896C22.485 18.9701 19.1434 22.2851 15.0377 22.2851H15.0347C13.6979 22.2844 12.3868 21.9282 11.2433 21.255L10.9712 21.0949L8.15184 21.8283L8.90461 19.1017V19.1018Z" fill="url(#paint1_linear_33_3093)"/>
+<path d="M12.7984 11.1732C12.6306 10.8032 12.454 10.7957 12.2945 10.7892C12.1639 10.7836 12.0145 10.7841 11.8653 10.7841C11.716 10.7841 11.4733 10.8398 11.2682 11.062C11.0629 11.2844 10.4844 11.8219 10.4844 12.9151C10.4844 14.0084 11.2869 15.0649 11.3987 15.2133C11.5107 15.3614 12.9479 17.6765 15.2241 18.5672C17.1158 19.3073 17.5007 19.1601 17.9113 19.123C18.3219 19.086 19.2362 18.5857 19.4228 18.0668C19.6095 17.548 19.6095 17.1033 19.5535 17.0104C19.4975 16.9178 19.3482 16.8622 19.1242 16.7511C18.9002 16.64 17.7993 16.1024 17.594 16.0283C17.3887 15.9542 17.2394 15.9172 17.0901 16.1397C16.9408 16.3618 16.5119 16.8622 16.3812 17.0104C16.2506 17.1589 16.1199 17.1774 15.896 17.0663C15.672 16.9548 14.9508 16.7204 14.0951 15.9635C13.4294 15.3746 12.98 14.6472 12.8493 14.4247C12.7187 14.2026 12.8353 14.0822 12.9476 13.9715C13.0483 13.8719 13.1716 13.712 13.2837 13.5823C13.3954 13.4525 13.4326 13.3599 13.5073 13.2118C13.5821 13.0634 13.5447 12.9337 13.4888 12.8225C13.4326 12.7114 12.9975 11.6126 12.7984 11.1732" fill="white"/>
+</g>
+<defs>
+<linearGradient id="paint0_linear_33_3093" x1="15.0008" y1="23.63" x2="15.0008" y2="6.31055" gradientUnits="userSpaceOnUse">
+<stop stop-color="#1FAF38"/>
+<stop offset="1" stop-color="#60D669"/>
+</linearGradient>
+<linearGradient id="paint1_linear_33_3093" x1="15" y1="23.9405" x2="15" y2="6" gradientUnits="userSpaceOnUse">
+<stop stop-color="#F9F9F9"/>
+<stop offset="1" stop-color="white"/>
+</linearGradient>
+<clipPath id="clip0_33_3093">
+<rect width="18" height="18" fill="white" transform="translate(6 6)"/>
+</clipPath>
+</defs>
+</svg>
+
+   
+  )
+}
+
+export function ReditIcon(){
+  return (
+    <svg width="30" height="30" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+<circle cx="15" cy="15" r="14.5" stroke="#FF4500"/>
+<g clip-path="url(#clip0_33_3099)">
+<path d="M15 23C19.4183 23 23 19.4183 23 15C23 10.5817 19.4183 7 15 7C10.5817 7 7 10.5817 7 15C7 19.4183 10.5817 23 15 23Z" fill="#FF4500"/>
+<path d="M20.3234 15.0762C20.3234 14.4277 19.799 13.9127 19.1599 13.9127C18.858 13.9123 18.5676 14.0285 18.3493 14.237C17.5481 13.6647 16.4512 13.2928 15.2305 13.2451L15.7645 10.7463L17.5003 11.1182C17.5195 11.5569 17.8818 11.9099 18.3302 11.9099C18.7879 11.9099 19.1599 11.538 19.1599 11.08C19.1599 10.6222 18.788 10.2503 18.3302 10.2503C18.0059 10.2503 17.7197 10.4411 17.5862 10.7177L15.6502 10.3076C15.5929 10.2979 15.5357 10.3076 15.4975 10.3362C15.4498 10.3648 15.4212 10.4124 15.4118 10.4697L14.8203 13.2546C13.5805 13.2928 12.4646 13.6647 11.6539 14.2466C11.4356 14.038 11.1452 13.9219 10.8433 13.9223C10.1947 13.9223 9.67969 14.4467 9.67969 15.0859C9.67969 15.5627 9.96575 15.9632 10.3664 16.1446C10.347 16.2612 10.3374 16.3792 10.3378 16.4974C10.3378 18.2903 12.4266 19.7497 15.0016 19.7497C17.5768 19.7497 19.6655 18.2999 19.6655 16.4974C19.6654 16.3792 19.6559 16.2612 19.6369 16.1446C20.0374 15.9632 20.3234 15.5531 20.3234 15.0762ZM12.3311 15.9059C12.3311 15.4482 12.7029 15.0762 13.1609 15.0762C13.6187 15.0762 13.9906 15.4481 13.9906 15.9059C13.9906 16.3637 13.6188 16.7358 13.1609 16.7358C12.703 16.7452 12.3311 16.3637 12.3311 15.9059ZM16.9759 18.1091C16.4036 18.6814 15.3163 18.7196 15.0016 18.7196C14.6774 18.7196 13.5901 18.6718 13.0273 18.1091C12.9416 18.0233 12.9416 17.8897 13.0273 17.8039C13.1131 17.7182 13.2466 17.7182 13.3325 17.8039C13.695 18.1664 14.458 18.2903 15.0016 18.2903C15.5453 18.2903 16.3177 18.1664 16.6706 17.8039C16.7565 17.7182 16.89 17.7182 16.9759 17.8039C17.0521 17.8897 17.0521 18.0233 16.9759 18.1091ZM16.8232 16.7452C16.3654 16.7452 15.9935 16.3734 15.9935 15.9156C15.9935 15.4577 16.3654 15.0859 16.8232 15.0859C17.2811 15.0859 17.653 15.4577 17.653 15.9156C17.653 16.3637 17.2811 16.7452 16.8232 16.7452Z" fill="white"/>
+</g>
+<defs>
+<clipPath id="clip0_33_3099">
+<rect width="16" height="16" fill="white" transform="translate(7 7)"/>
+</clipPath>
+</defs>
+</svg>
+
+
+  )
+}
+
+
+export function MealRequestIcon(props) {
+  return (
+    <svg
+      width="15"
+      height="15"
+      viewBox="0 0 15 15"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M4.20156 0.59961C3.9533 0.599479 3.71111 0.676352 3.50837 0.819632C3.30563 0.962912 3.15232 1.16555 3.06956 1.39961H1.80156C1.4833 1.39961 1.17808 1.52604 0.953034 1.75108C0.727991 1.97613 0.601563 2.28135 0.601562 2.59961V12.1996C0.601562 12.5179 0.727991 12.8231 0.953034 13.0481C1.17808 13.2732 1.4833 13.3996 1.80156 13.3996H7.20716C6.69763 12.7822 6.3662 12.0374 6.24858 11.2456C6.13095 10.4537 6.23158 9.6448 6.53963 8.90591C6.84768 8.16703 7.35147 7.52618 7.99675 7.05239C8.64203 6.57861 9.40433 6.28985 10.2016 6.21721V2.59961C10.2016 2.28135 10.0751 1.97613 9.85009 1.75108C9.62505 1.52604 9.31982 1.39961 9.00156 1.39961H7.73356C7.65081 1.16555 7.4975 0.962912 7.29475 0.819632C7.09201 0.676352 6.84982 0.599479 6.60156 0.59961H4.20156ZM3.80156 1.79961C3.80156 1.69352 3.84371 1.59178 3.91872 1.51677C3.99373 1.44175 4.09548 1.39961 4.20156 1.39961H6.60156C6.70765 1.39961 6.80939 1.44175 6.88441 1.51677C6.95942 1.59178 7.00156 1.69352 7.00156 1.79961C7.00156 1.9057 6.95942 2.00744 6.88441 2.08245C6.80939 2.15747 6.70765 2.19961 6.60156 2.19961H4.20156C4.09548 2.19961 3.99373 2.15747 3.91872 2.08245C3.84371 2.00744 3.80156 1.9057 3.80156 1.79961ZM7.68476 6.88281L4.88476 9.68281C4.84761 9.72006 4.80347 9.74961 4.75487 9.76978C4.70627 9.78994 4.65418 9.80032 4.60156 9.80032C4.54895 9.80032 4.49685 9.78994 4.44826 9.76978C4.39966 9.74961 4.35552 9.72006 4.31836 9.68281L3.11836 8.48281C3.08117 8.44562 3.05167 8.40147 3.03154 8.35288C3.01142 8.30428 3.00106 8.2522 3.00106 8.19961C3.00106 8.14701 3.01142 8.09493 3.03154 8.04634C3.05167 7.99775 3.08117 7.9536 3.11836 7.91641C3.19347 7.8413 3.29534 7.7991 3.40156 7.7991C3.45416 7.7991 3.50624 7.80946 3.55483 7.82959C3.60342 7.84972 3.64757 7.87922 3.68476 7.91641L4.60156 8.83401L7.11836 6.31641C7.19347 6.2413 7.29534 6.1991 7.40156 6.1991C7.50778 6.1991 7.60965 6.2413 7.68476 6.31641C7.75987 6.39152 7.80207 6.49339 7.80207 6.59961C7.80207 6.70583 7.75987 6.8077 7.68476 6.88281ZM14.2016 10.5996C14.2016 11.5544 13.8223 12.4701 13.1471 13.1452C12.472 13.8203 11.5563 14.1996 10.6016 14.1996C9.64678 14.1996 8.73111 13.8203 8.05598 13.1452C7.38085 12.4701 7.00156 11.5544 7.00156 10.5996C7.00156 9.64483 7.38085 8.72916 8.05598 8.05402C8.73111 7.37889 9.64678 6.99961 10.6016 6.99961C11.5563 6.99961 12.472 7.37889 13.1471 8.05402C13.8223 8.72916 14.2016 9.64483 14.2016 10.5996ZM11.0016 8.99961C11.0016 8.89352 10.9594 8.79178 10.8844 8.71677C10.8094 8.64175 10.7076 8.59961 10.6016 8.59961C10.4955 8.59961 10.3937 8.64175 10.3187 8.71677C10.2437 8.79178 10.2016 8.89352 10.2016 8.99961V10.1996H9.00156C8.89548 10.1996 8.79373 10.2418 8.71872 10.3168C8.6437 10.3918 8.60156 10.4935 8.60156 10.5996C8.60156 10.7057 8.6437 10.8074 8.71872 10.8825C8.79373 10.9575 8.89548 10.9996 9.00156 10.9996H10.2016V12.1996C10.2016 12.3057 10.2437 12.4074 10.3187 12.4825C10.3937 12.5575 10.4955 12.5996 10.6016 12.5996C10.7076 12.5996 10.8094 12.5575 10.8844 12.4825C10.9594 12.4074 11.0016 12.3057 11.0016 12.1996V10.9996H12.2016C12.3076 10.9996 12.4094 10.9575 12.4844 10.8825C12.5594 10.8074 12.6016 10.7057 12.6016 10.5996C12.6016 10.4935 12.5594 10.3918 12.4844 10.3168C12.4094 10.2418 12.3076 10.1996 12.2016 10.1996H11.0016V8.99961Z"
+        fill="white"
+      />
+    </svg>
+  );
+}
 export function CallIcon(props) {
   return (
     <svg
