@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import styles from "./store.module.css";
 
 import Head from "next/head";
@@ -17,11 +17,12 @@ import sale from "../../../public/assets/store_pics/sale.jpg";
 import { FormModal } from "../modal/form-modal";
 import { SuccessModal } from "../suggest-store/success-modal";
 import { useRouter } from "next/navigation";
+import axios from "../../util/Api";
 
 function Store(props) {
   const [openModal, setOpenModal] = useState(false);
   const [openSuccessModal, setOpenSuccessModal] = useState(false);
-  const [selectItem, setSelectItem] = useState({});
+  const [selectedStore, setSelectedStore] = useState([])
   const router = useRouter();
 
   function handleSearch(e) {
@@ -45,14 +46,37 @@ function Store(props) {
     //   })
     // }
   }
-  console.log(props.store, "props11");
+  console.log(props?.item, "props11");
   const filteredItem = () => {
     return props?.items?.filter((data) => data?.item_type === "Meal");
   };
   const filteredProducts = () => {
     return props?.items?.filter((data) => data?.item_type === "Product");
   };
-  console.log(filteredProducts(), "filtered");
+
+  const fetchOneStore = async (storeId) => {
+
+    try {
+      const response = await axios.get(`/inventory/get-store-inventory/${storeId}`);
+      setSelectedStore(response.data.data);
+
+    } catch (error) {
+      console.error(`Attempt ${i + 1}: Error fetching store inventory`, error);
+      if (i === retries - 1) {
+        console.error("All retry attempts failed");
+      }
+    }
+
+  };
+
+  useEffect(() => {
+    if (props?.store?._id) {
+      fetchOneStore(props.store._id);
+    } else {
+      console.error("Store ID is undefined");
+    }
+  }, [props?.store?._id]);
+  console.log(props.store, 'props?.store?._id')
   return (
     <>
       <Head>
@@ -86,7 +110,11 @@ function Store(props) {
         <div className={styles.product_section_2}>
           <div className={styles.product_section_2_col_1}>
             <Image
-              src={props.store?.profile_picture}
+              src={
+                props.store?.profile_picture
+                  ? props.store?.profile_picture
+                  : "/assets/store_pics/no-image-store.png"
+              }
               alt="pop up"
               className={styles.product_section_2_main_img}
               height={350}
@@ -198,55 +226,62 @@ function Store(props) {
             </div>
             <div className={styles.productcard_col_2}>
               <div className={styles.productcard_productcards}>
-                {filteredItem()
-                  ?.slice(0, 6)
-                  ?.map((data, index) => {
+                {selectedStore?.inventory
+                  ?.filter((elem) => elem.item_type === 'Meal')
+                  .map((meal, idx) => {
                     return (
-                      <div
-                        key={index}
-                        className={styles.productcard_productcard}
-                      >
-                        <div
-                          className={
-                            styles.productcard_productcard_img_container
-                          }
-                        >
+                      <div key={idx}
+                        className={styles.productcard_productcard}>
+                        <div className={
+                          styles.productcard_productcard_img_container
+                        }>
                           <img
-                            priority
-                            src={data?.itemImage0}
-                            alt="Store"
+                            src={
+                              meal?.item?.itemImage0
+                                ? meal.item.itemImage0
+                                : !meal?.item?.itemImage0 &&
+                                  meal?.item?.item_type === "Meal"
+                                  ? "/assets/store_pics/no-image-meal.png"
+                                  : !meal?.item?.itemImage0 && meal?.item?.item_type === "Product"
+                                    ? "/assets/store_pics/no-image-product.png"
+                                    : !meal?.item?.itemImage0 && meal?.item?.item_type === "Utensil"
+                                      ? "/assets/store_pics/no-image-utensil.png"
+                                      : ""
+                            }
+                            alt=""
                             className={styles.productcard_productcard_img}
                           />
                         </div>
+
                         <div className={styles.productcard_productcard_col}>
-                          <h6 className={styles.productcard_productcard_name}>
-                            {data.item_name}
-                          </h6>
-                          <p className={styles.productcard_productcard_price}>
-                           {data.item_price ?  props?.store?.currency?.symbol + data.item_price : "N/A"}
-                          </p>
+                          <p className={styles.productcard_productcard_name}>{meal?.item?.item_name}</p>
+
                         </div>
                         <div className={styles.productcard_productcard_col}>
-                          <div className={styles.rate}>
-                            {Array(5)
-                              .fill("_")
-                              .map((_, idx) => (
-                                <GrStar
-                                  size={20}
-                                  key={idx + _}
-                                  color={
-                                    data.average_rating > idx
-                                      ? "#04D505"
-                                      : "rgba(0,0,0,0.5)"
-                                  }
-                                />
-                              ))}
+                          <div className={styles.store_review_rating_icons}>
+                            <div className={styles.rate}>
+                              {Array(5)
+                                .fill("_")
+                                .map((_, idx) => (
+                                  <GrStar
+                                    size={20}
+                                    key={idx + _}
+                                    color={
+                                      meal?.average_rating > idx
+                                        ? "#04D505"
+                                        : "rgba(0,0,0,0.5)"
+                                    }
+                                  />
+                                ))}
+                            </div>
                           </div>
-
-                          <p
-                            className={styles.productcard_productcard_duration}
-                          >
-                            {data.meal_cook_time ? data.meal_cook_time : 0} min
+                          <p className={styles.name2}>
+                            {meal?.item?.item_price
+                              ? meal?.storeId.filter((elem) => elem.store_name === props?.store?.store_name)
+                                .map((elem) => (
+                                  `${elem.currency?.symbol} ${meal.item.item_price}`
+                                ))
+                              : "N/A"}
                           </p>
                         </div>
                       </div>
@@ -262,30 +297,36 @@ function Store(props) {
             </div>
             <div className={styles.productcard_col_2}>
               <div className={styles.productcard_productcards}>
-                {filteredProducts()
-                  ?.slice(0, 6)
-                  ?.map((data, index) => {
+                {selectedStore?.inventory
+                  ?.filter((elem) => elem.item_type === 'Product')
+                  .map((meal, idx) => {
                     return (
-                      <div
-                        key={index}
-                        className={styles.productcard_productcard}
-                      >
-                        <div
-                          className={
-                            styles.productcard_productcard_img_container
-                          }
-                        >
+                      <div key={idx}
+                        className={styles.productcard_productcard}>
+                        <div className={
+                          styles.productcard_productcard_img_container
+                        }>
                           <img
-                            priority
-                            src={data.itemImage0}
-                            alt="Store"
+                            src={
+                              meal?.item?.itemImage0
+                                ? meal.item.itemImage0
+                                : !meal?.item?.itemImage0 &&
+                                  meal?.item?.item_type === "Meal"
+                                  ? "/assets/store_pics/no-image-meal.png"
+                                  : !meal?.item?.itemImage0 && meal?.item?.item_type === "Product"
+                                    ? "/assets/store_pics/no-image-product.png"
+                                    : !meal?.item?.itemImage0 && meal?.item?.item_type === "Utensil"
+                                      ? "/assets/store_pics/no-image-utensil.png"
+                                      : ""
+                            }
+                            alt=""
                             className={styles.productcard_productcard_img}
                           />
                         </div>
+
                         <div className={styles.productcard_productcard_col}>
-                          <h6 className={styles.productcard_productcard_name}>
-                            {data.item_name}
-                          </h6>
+                          <p className={styles.productcard_productcard_name}>{meal?.item?.item_name}</p>
+
                         </div>
                         <div className={styles.productcard_productcard_col}>
                           <div className={styles.store_review_rating_icons}>
@@ -297,7 +338,7 @@ function Store(props) {
                                     size={20}
                                     key={idx + _}
                                     color={
-                                      data.average_rating > idx
+                                      meal?.average_rating > idx
                                         ? "#04D505"
                                         : "rgba(0,0,0,0.5)"
                                     }
@@ -305,8 +346,13 @@ function Store(props) {
                                 ))}
                             </div>
                           </div>
-                          <p className={styles.productcard_productcard_price}>
-                          {data.item_price ?  props?.store?.currency?.symbol + data.item_price : "N/A"}
+                          <p className={styles.name2}>
+                            {meal?.item?.item_price
+                              ? meal?.storeId.filter((elem) => elem.store_name === props?.store?.store_name)
+                                .map((elem) => (
+                                  `${elem.currency?.symbol} ${meal.item.item_price}`
+                                ))
+                              : "N/A"}
                           </p>
                         </div>
                       </div>
