@@ -9,23 +9,24 @@ import { IndividualModal } from "../modal/individual-meal-product";
 import { useMediaQuery } from "../../hooks/usemediaquery";
 import { Mealmodal } from "../mobile/meal-modal";
 import { Element, scroller } from "react-scroll";
-import { ScrollableElement } from "../smooth-scroll-link";
+import { addToCart } from "../../actions";
+import { useDispatch } from "react-redux";
+import { canItemBeAddedToCart } from "../../util/canAddToCart";
 
 export const PopularMeals = () => {
   const matches = useMediaQuery("(min-width: 920px)");
   const [meals, setMeals] = useState([]);
-  const [visibleMeals, setVisibleMeals] = useState(8);
   const [selectedItem, setSelectedItem] = useState({});
   const [selectGrocery, setSelectGrocery] = useState([]);
   const [openModal, setOpenModal] = useState(false);
   const [show, setShow] = useState(false);
   const [openList, setOpenList] = useState(false);
   const [quantity, setQuantity] = useState(0);
+  const [serve, setServe] = useState(0);
+  const dispatch = useDispatch();
   const ref = useRef(null);
+  const [selectedItemId, setSelectedItemId] = useState(null);
 
-  const loadMore = () => {
-    setVisibleMeals(visibleMeals + 4);
-  };
   const router = useRouter();
   const [itemToAdd, setItemAdd] = useState({
     listName: "",
@@ -44,7 +45,6 @@ export const PopularMeals = () => {
       },
     };
 
-    console.log(payload, "payload");
     try {
       const response = await axios(`/groceries`, {
         method: "post",
@@ -57,30 +57,87 @@ export const PopularMeals = () => {
       console.log(error);
     }
   };
+
+  const addItemToCart = (item, qty) => {
+    const user = JSON.parse(localStorage.getItem("user"));
+    let canAddToCart = canItemBeAddedToCart(item);
+
+    if (qty == 0) {
+      toast.error("Pls add a quantity");
+    } else {
+      if (canAddToCart) {
+        const payload = {
+          userId: user && user._id ? user._id : "",
+          storeId: "",
+          store_name: "",
+          itemId: item._id,
+          quantity: qty,
+          item_price: item.item_price,
+          currency: "$",
+          item_image: item.itemImage0,
+          itemName: item.item_name,
+          item_type: item.item_type ? item.item_type : "Meal",
+        };
+        try {
+          dispatch(addToCart(payload));
+          setOpenList(false);
+          setShow(false);
+          setOpenModal(false);
+        } catch (error) {
+          console.log(error);
+        }
+      }
+    }
+  };
   const [details, setDetails] = useState({
     listName: "",
     description: "",
     id: "",
     status: "",
   });
+  const [currentPage, setCurrentPage] = useState(1);
+  const [hasMoreData, setHasMoreData] = useState(true);
 
-  const fetchMeals = async () => {
+  const [uniqueItemIds, setUniqueItemIds] = useState(new Set());
+
+  const fetchMeals = async (page) => {
     try {
-      const response = await axios(`/items/1?type=Meal&status=all&limit=50`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      console.log(response.data.data.items, "ressw");
-      setMeals(response.data.data.items);
+      const response = await axios(
+        `/items/${
+          page ? page : currentPage
+        }?type=Meal&status=Public&limit=4&average_rating=1`,
+        {
+          method: "GET",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+      const totalItems = response.data.data.count;
+      const allItems = response.data.data.items;
+
+      // const filteredItems = allItems.filter((meal) => meal.average_rating);
+
+      const newItems = allItems.filter((item) => !uniqueItemIds.has(item._id));
+
+      setMeals((prev) => [...prev, ...allItems]);
+      setUniqueItemIds(
+        new Set([...uniqueItemIds, ...newItems.map((item) => item._id)])
+      );
+
+      setHasMoreData(totalItems > currentPage * 8);
     } catch (error) {
       console.log(error);
     }
   };
+
+  const loadMore = async () => {
+    setCurrentPage(currentPage + 1);
+    await fetchMeals(currentPage + 1);
+  };
   useEffect(() => {
     fetchMeals();
-  }, []);
+  }, [currentPage]);
   const fetchGroceryList = async () => {
     try {
       const response = await axios(`/groceries/list`, {
@@ -89,35 +146,29 @@ export const PopularMeals = () => {
           "Content-Type": "application/json",
         },
       });
-      console.log(response.data.data.data, "groceries");
       setSelectGrocery(response.data.data.data);
     } catch (error) {}
   };
   useEffect(() => {
     fetchGroceryList();
   }, []);
-  const filteredMeals = meals.filter(
-    (meal) => meal.item_type === "Meal" && meal.average_rating
-  );
-  console.log(filteredMeals, "fill");
+  const filteredMeals = meals.filter((meal) => meal.average_rating);
 
   useEffect(() => {
-    // Get the hash value from the URL
     const hash = window.location.hash;
 
-    // Use the hash value as the target ID for scrolling
-    const targetId = hash ? hash.substring(1) : 'store';
+    const targetId = hash ? hash.substring(1) : "store";
 
-    // Scroll to the target section
     if (targetId) {
       scroller.scrollTo(targetId, {
         duration: 1000,
         delay: 0,
-        smooth: 'easeInOutQuart',
+        smooth: "easeInOutQuart",
         offset: -1000,
       });
     }
   }, []);
+  console.log(selectedItem, "selectedItem");
   return (
     <div className={styles.mealContainer}>
       <Element
@@ -128,56 +179,63 @@ export const PopularMeals = () => {
         Popular Meals
       </Element>
       <div className={styles.stores2}>
-        {filteredMeals
-          .slice(0, visibleMeals)
-          .filter((meal) => Boolean(meal.total_rating))
-          .map((meal, idx) => {
-            return (
-              <div
-                className={styles.card1}
-                key={idx}
-                onClick={() => {
-                  setSelectedItem(meal);
-                  setOpenModal(true);
-                }}
-              >
-                {
-                  <div className={styles.box}>
-                    <img src={meal?.itemImage0} className={styles.storeImg1} />
-                    <div className={styles.flex}>
-                      <p className={styles.name2}>{meal.item_name}</p>
-                      <p>${meal.item_price ? meal.item_price : "0"}</p>
-                    </div>
-                    <p className={styles.storeName}>Chop Chow Official Store</p>
-                    <div className={styles.flex}>
-                      <div>
-                        {Array(5)
-                          .fill("_")
-                          .map((_, idx) => (
-                            <GoStarFill
-                              key={idx + _}
-                              color={
-                                meal.average_rating > idx
-                                  ? "#04D505"
-                                  : "rgba(0,0,0,0.5)"
-                              }
-                              style={{ marginLeft: ".2rem" }}
-                            />
-                          ))}
-                      </div>
-                      <p className={styles.prep}> 0 mins </p>
-                    </div>
+        {meals.map((meal, idx) => {
+          return (
+            <div
+              className={styles.card1}
+              key={meal._id}
+              onClick={() => {
+                setSelectedItem(meal);
+                setOpenModal(true);
+                setSelectedItemId(meal._id);
+                console.log(selectedItem, "slected");
+              }}
+            >
+              {
+                <div className={styles.box}>
+                  <img
+                    src={
+                      meal?.itemImage0
+                        ? meal?.itemImage0
+                        : "/assets/store_pics/no-image-meal.png"
+                    }
+                    className={styles.storeImg1}
+                  />
+                  <div className={styles.flex}>
+                    <p className={styles.name2}>{meal.item_name}</p>
+                    <p>${meal.item_price ? meal.item_price : "0"}</p>
                   </div>
-                }
-              </div>
-            );
-          })}
+                  <p className={styles.storeName}>Chop Chow Official Store</p>
+                  <div className={styles.flex}>
+                    <div>
+                      {Array(5)
+                        .fill("_")
+                        .map((_, idx) => (
+                          <GoStarFill
+                            key={idx + _}
+                            color={
+                              meal.average_rating > idx
+                                ? "#04D505"
+                                : "rgba(0,0,0,0.5)"
+                            }
+                            style={{ marginLeft: ".2rem" }}
+                          />
+                        ))}
+                    </div>
+                    <p className={styles.prep}> 0 mins </p>
+                  </div>
+                </div>
+              }
+            </div>
+          );
+        })}
         {!matches ? (
           <Mealmodal
             openList={openList}
             openModal={openModal}
             selectGrocery={selectGrocery}
             selectedItem={selectedItem}
+            selectedItemId={selectedItemId}
             setOpenList={setOpenList}
             setOpenModal={setOpenModal}
             show={show}
@@ -188,6 +246,9 @@ export const PopularMeals = () => {
             setQuantity={setQuantity}
             quantity={quantity}
             setShow={setShow}
+            addToCart={addItemToCart}
+            serve={serve}
+            setServe={setServe}
           />
         ) : (
           <IndividualModal
@@ -204,11 +265,15 @@ export const PopularMeals = () => {
             setItemAdd={setItemAdd}
             setQuantity={setQuantity}
             quantity={quantity}
+            selectedItemId={selectedItemId}
             setShow={setShow}
+            addToCart={addItemToCart}
+            serve={serve}
+            setServe={setServe}
           />
         )}
       </div>
-      <p className={styles.view} onClick={() => loadMore()}>
+      <p className={styles.view} onClick={hasMoreData ? loadMore : () => {}}>
         View More
       </p>
       <div className={styles.border} />
