@@ -19,51 +19,53 @@ import { toast } from "react-toastify";
 import { base_url } from "../../util/Api";
 import { withRouter } from "next/router";
 
+const initialState = {
+  utensilName: "",
+  utensilImage1: "",
+  utensilImage2: "",
+  utensilImage3: "",
+  utensilImage4: "",
+  utensilImagesData: [],
+  intro: "",
+
+  sizeNames: [],
+  // do we need product group list AND strings ?
+  descriptionGroupList: [],
+  // store product names of inputted strings to compare with db products
+  descriptionStrings: [],
+  // do we want to use current ingredient formats ? Yes.
+  currentIngredient: "",
+  measurement: "",
+  quantity: "",
+  currentProductImgSrc: null,
+  currentProductDisplayIndex: 0,
+
+  currentStore: "",
+
+  // we need to update how we create image paths
+  productImg_path: "",
+  suggested_stores: [],
+  currProductIndexInDBsProductsList: -1,
+  // currStoreIndexIfExistsInProductsList: -1,
+
+  instructionWordlength: 0,
+
+  suggestedCategories: [],
+
+  booleanOfDisplayOfDialogBoxConfirmation: false,
+
+  //mealsModal controller
+  openModal: false,
+  stepInputs: [],
+};
+
 class SuggestKitchenUtensilForm extends Component {
   utensilsList = [];
   ingredientsQuantityMeasurements = [];
 
   constructor(props) {
     super(props);
-    this.state = {
-      utensilName: "",
-      utensilImage1: "",
-      utensilImage2: "",
-      utensilImage3: "",
-      utensilImage4: "",
-      utensilImagesData: [],
-      intro: "",
-
-      sizeNames: [],
-      // do we need product group list AND strings ?
-      descriptionGroupList: [],
-      // store product names of inputted strings to compare with db products
-      descriptionStrings: [],
-      // do we want to use current ingredient formats ? Yes.
-      currentIngredient: "",
-      measurement: "",
-      quantity: "",
-      currentProductImgSrc: null,
-      currentProductDisplayIndex: 0,
-
-      currentStore: "",
-
-      // we need to update how we create image paths
-      productImg_path: "",
-      suggested_stores: [],
-      currProductIndexInDBsProductsList: -1,
-      // currStoreIndexIfExistsInProductsList: -1,
-
-      instructionWordlength: 0,
-
-      suggestedCategories: [],
-
-      booleanOfDisplayOfDialogBoxConfirmation: false,
-
-      //mealsModal controller
-      openModal: false,
-      stepInputs: [],
-    };
+    this.state = initialState;
 
     this.closeModal = this.closeModal.bind(this);
     // this.handleStoreNameInput = this.handleStoreNameInput.bind(this);
@@ -142,7 +144,7 @@ class SuggestKitchenUtensilForm extends Component {
     let doc = document.querySelector("#formutensil");
     if (doc) {
       setInterval(() => {
-        localStorage.setItem("suggestUtensilForm", JSON.stringify(this.state));
+        // localStorage.setItem("suggestUtensilForm", JSON.stringify(this.state));
       }, 100);
     }
 
@@ -150,8 +152,8 @@ class SuggestKitchenUtensilForm extends Component {
       let {
         utensilName,
         intro,
-
         sizeNames,
+        utensilImagesData,
         // do we need product group list AND strings ?
         descriptionGroupList,
         // store product names of inputted strings to compare with db products
@@ -172,7 +174,7 @@ class SuggestKitchenUtensilForm extends Component {
         // currStoreIndexIfExistsInProductsList,
 
         suggestedCategories,
-
+        productImagesData,
         booleanOfDisplayOfDialogBoxConfirmation,
       } = JSON.parse(localStorage.getItem("suggestUtensilForm"));
 
@@ -181,9 +183,10 @@ class SuggestKitchenUtensilForm extends Component {
         utensilImage: "",
         utensilImageName: "",
         utensilImageData: "",
-        utensilImagesData: [],
+        utensilImagesData: Array.isArray(productImagesData)
+          ? productImagesData
+          : [],
         intro,
-
         sizeNames,
         // do we need product group list AND strings ?
         descriptionGroupList,
@@ -351,7 +354,7 @@ class SuggestKitchenUtensilForm extends Component {
         searchResult.indexOf(true);
       console.log(
         "Curr Product Index If Exists In Products List is: \n" +
-          tmpcurrMeasurementIndexInDBsMeasurementList
+        tmpcurrMeasurementIndexInDBsMeasurementList
       );
 
       // check if product name is an existing product
@@ -397,10 +400,11 @@ class SuggestKitchenUtensilForm extends Component {
     // descriptionObject.calories = 0;
 
     console.log(descriptionObject, "descriptionObjectdescriptionObject");
+    // console.log(this.state?.descriptionObject, 'descree')
 
     this.setState({
       descriptionGroupList: [
-        ...this.state?.descriptionObject,
+        // ...this.state?.descriptionObject,
         descriptionObject,
       ],
     });
@@ -413,10 +417,10 @@ class SuggestKitchenUtensilForm extends Component {
   };
 
   handleAddDescriptionChip(chip) {
-    this.setState({
-      descriptionStrings: [...this.state.descriptionStrings, chip],
-    });
-  }
+    this.setState(prevState => ({
+        descriptionStrings: [...(prevState.descriptionStrings || []), chip],
+    }));
+}
 
   handleDeleteSizeChip(chip) {
     var array = this.state.descriptionStrings; // make a separate copy of the array
@@ -546,7 +550,17 @@ class SuggestKitchenUtensilForm extends Component {
     }
     if (utensilImage4) {
       suggestProductForm.append("item_images", utensilImage4);
+
     }
+
+    if (!utensilImage1 && !utensilImage2 && !utensilImage3 && !utensilImage4) {
+      const img = await fetch(
+        "/assets/store_pics/no-image-utensil.png"
+      );
+      const blob = await img.blob()
+      suggestProductForm.append("item_images", blob);
+    }
+
     // descriptionGroupList.map((individualDescriptions) => {
     //   console.log(individualDescriptions);
     //   suggestProductForm.append('product_descriptions', JSON.stringify(individualDescriptions));
@@ -563,9 +577,8 @@ class SuggestKitchenUtensilForm extends Component {
         object_name: ele.description_name,
         object_quantity: ele.quantity,
         object_measurement: ele.measurement,
-        formatted_string: `${this.capitalizeWords(ele.description_name)} : ${
-          ele.quantity
-        }${ele.measurement}`,
+        formatted_string: `${this.capitalizeWords(ele.description_name)} : ${ele.quantity
+          }${ele.measurement}`,
       };
     });
 
@@ -575,23 +588,21 @@ class SuggestKitchenUtensilForm extends Component {
       let qty = "";
 
       const ele = splited[1];
-      for (let i = 0; i < ele.length; i++) {        
-        if (!Object.is(Number(ele.charAt(i)), NaN)){
-          console.log(ele.charAt(i), 'ele.charAt(i)')
-          qty = qty.concat(ele.charAt(i))
+      for (let i = 0; i < ele.length; i++) {
+        if (!Object.is(Number(ele.charAt(i)), NaN)) {
+          console.log(ele.charAt(i), "ele.charAt(i)");
+          qty = qty.concat(ele.charAt(i));
         }
       }
       const measurement = ele.slice(ele.indexOf(qty));
-      const formatted_string = `${this.capitalizeWords(
-        descripName
-      )} : ${ele}`;
+      const formatted_string = `${this.capitalizeWords(descripName)} : ${ele}`;
 
       return {
         object_name: descripName,
         object_quantity: Number(qty),
         object_measurement: measurement,
-        formatted_string
-      }
+        formatted_string,
+      };
     });
 
     suggestProductForm.append("description", JSON.stringify(arr_2));
@@ -660,6 +671,8 @@ class SuggestKitchenUtensilForm extends Component {
           console.log(response);
           console.log("Display Product submitted successfully");
           toast.success("Kitchen Utensils submitted successfully");
+          localStorage.setItem("suggestUtensilForm", JSON.stringify({}));
+          this.setState(initialState);
           // window.location.href = "/SuggestProduct"
         } else {
           console.log("Something wrong happened ");
@@ -774,7 +787,12 @@ class SuggestKitchenUtensilForm extends Component {
             </div>
           </div>
           <h3>Utensil Descriptions</h3>
-          <div className={styles.suggestion_form}>
+          <div className={styles.suggestion_form} onKeyDown={(event) => {
+            if (event.key === "Enter") {
+              event.preventDefault();
+              this.addDescription(event)
+            }
+          }}>
             <div className={styles.suggestion_form_group}>
               <label
                 htmlFor="descriptionName"
@@ -800,6 +818,7 @@ class SuggestKitchenUtensilForm extends Component {
                     Quantity
                   </label>
                   <TextField
+                    inputProps={{ min: 0 }}
                     fullWidth
                     id="quantity"
                     type="number"
@@ -821,7 +840,7 @@ class SuggestKitchenUtensilForm extends Component {
                   </label>
                   <Autocomplete
                     id="measurement"
-                    options={this.props.measurements.map((option) => option)}
+                    options={this.props.measurements?.map((option) => option)}
                     value={this.state.measurement}
                     onChange={this.handleMeasurement}
                     freeSolo
@@ -879,15 +898,9 @@ class SuggestKitchenUtensilForm extends Component {
                   freeSolo
                   clearOnBlur
                   onBlur={this.categoryBlur}
-                  // filterSelectedOptions
-                  options={this.props.categories?.map((option) => option)}
-                  // onChange={(ev,val)=>this.handleCategoryDropdownChange(ev,val)}
-                  onChange={(e, newValue) =>
-                    this.handleCategoryDropdownChange(newValue)
-                  }
-                  // getOptionLabel={option => option}
-                  // renderTags={() => {}}
-                  value={this.state.suggestedCategories}
+                  options={this.props.categories ? this.props.categories : []}
+                  onChange={(e, newValue) => this.handleCategoryDropdownChange(newValue)}
+                  value={Array.isArray(this.state.suggestedCategories) ? this.state.suggestedCategories : []}
                   renderInput={(params) => (
                     <TextField
                       {...params}
@@ -897,6 +910,7 @@ class SuggestKitchenUtensilForm extends Component {
                     />
                   )}
                 />
+
                 <Button
                   variant="contained"
                   disableRipple
@@ -910,7 +924,7 @@ class SuggestKitchenUtensilForm extends Component {
               </div>
             </div>
             <Stack direction="row" spacing={1} className={styles.stack}>
-              {this.state.suggestedCategories.map((data, index) => (
+              {this.state.suggestedCategories?.map((data, index) => (
                 <Chip
                   key={index}
                   label={data}
