@@ -16,9 +16,12 @@ import Image from "next/image";
 import { AiOutlineClose } from "react-icons/ai";
 import { toast } from "react-toastify";
 import { base_url } from "../../util/Api";
+import ScannerImg from "../../../public/assets/home/scanner.svg"
+import { IoIosCloseCircle } from "react-icons/io";
 
 class SuggestProductForm extends Component {
   allProductNames = [];
+  ref = React.createRef()
   productNames = [
     "Spinach",
     "Brown Beans",
@@ -91,6 +94,7 @@ class SuggestProductForm extends Component {
       productImage3: "",
       productImage4: "",
       productImagesData: [],
+      imageData: [],
       productDescription: "",
       intro: "",
       ingredientsInItem: [],
@@ -142,6 +146,15 @@ class SuggestProductForm extends Component {
       //productsModal controller
       openModal: false,
       stepInputs: [],
+      measurements: [],
+
+      //Content extraction
+      image: {
+        file: [],
+        url: [],
+        loading: false,
+        visible: false
+      },
     };
 
     this.closeModal = this.closeModal.bind(this);
@@ -153,7 +166,7 @@ class SuggestProductForm extends Component {
   ///////////////////////////////////////////////////////////////////////////////////////
   componentDidMount() {
     console.log("nutritional-strings=-=-=-", this.state);
-    console.log(this.state.ingredientGroupList, 'ingredientGroupList')
+    console.log(this.state.ingredientGroupList, "ingredientGroupList");
     // get all product Names***
     // var url = "https://chopchowdev/api/products/getAllProducts";
     var url = `${base_url}/products/getAllProducts`;
@@ -215,6 +228,28 @@ class SuggestProductForm extends Component {
     //   .catch((err) => {
     //     console.log(err);
     //   });
+    const getAllMeasurement = async (newPage) => {
+      try {
+        const resp = await axios.get(
+          `/measurement/1?withPaginate=false&status=Public`
+        );
+        if (
+          Array.isArray(resp?.data?.data?.measurement) &&
+          resp?.data?.data?.measurement?.length
+        ) {
+          const mesr_names = resp.data.data?.measurement.map(
+            (ele) => ele.measurement_name
+          );
+          this.setState({
+            ...this.state,
+            measurements: mesr_names,
+          });
+        }
+      } catch (e) {
+        console.log("err-", e);
+      }
+    };
+    getAllMeasurement();
     this.categories = this.categories;
     if (typeof window !== "undefined") {
       let doc = document.querySelector("#formproduct");
@@ -241,8 +276,10 @@ class SuggestProductForm extends Component {
       if (localStorage.getItem("suggestProductForm")) {
         let {
           productName,
+          imageData,
           intro,
           productDescription = intro,
+          productImagesData,
           ingredientsInItem,
           ingredientNames,
           item_description,
@@ -252,7 +289,7 @@ class SuggestProductForm extends Component {
           sizeGroupList,
           // store product names of inputted strings to compare with db products
           ingredientStrings = ingredientsInItem,
-          nutritionalStrings= item_description,
+          nutritionalStrings = item_description,
           sizeStrings,
           // do we want to use current ingredient formats ? Yes.
           currentIngredient,
@@ -285,6 +322,8 @@ class SuggestProductForm extends Component {
           productName,
           productDescription,
           intro,
+          productImagesData,
+          imageData,
           ingredientsInItem,
           item_description,
           ingredientNames,
@@ -324,6 +363,7 @@ class SuggestProductForm extends Component {
         });
       }
     }
+
   }
 
   onInputChange = (e) => {
@@ -449,7 +489,7 @@ class SuggestProductForm extends Component {
       const tmpcurrProductIndexInDBsProductsList = searchResult.indexOf(true);
       console.log(
         "Curr Product Index If Exists In Products List is: \n" +
-          tmpcurrProductIndexInDBsProductsList
+        tmpcurrProductIndexInDBsProductsList
       );
 
       // check if product name is an existing product
@@ -493,7 +533,7 @@ class SuggestProductForm extends Component {
         searchResult.indexOf(true);
       console.log(
         "Curr Product Index If Exists In Products List is: \n" +
-          tmpcurrMeasurementIndexInDBsMeasurementList
+        tmpcurrMeasurementIndexInDBsMeasurementList
       );
 
       // check if product name is an existing product
@@ -535,7 +575,7 @@ class SuggestProductForm extends Component {
         searchResult.indexOf(true);
       console.log(
         "Curr Product Index If Exists In Products List is: \n" +
-          tmpcurrMeasurementIndexInDBsMeasurementList
+        tmpcurrMeasurementIndexInDBsMeasurementList
       );
 
       // check if product name is an existing product
@@ -705,17 +745,16 @@ class SuggestProductForm extends Component {
   ) => {
     console.log(
       "COMES IN addIngredientToProduct",
-      this.state.nutritionalStrings
+      this.state.currentIngredientQuantity
     );
     event.preventDefault();
     var properIngredientStringSyntax;
-    var ingredientValue = document.getElementById(id).value;
-    var quantityValue = document.getElementById(qty).value;
+    var ingredientValue = document.getElementById(id).value.split(' ').map((ele) => ele.slice(0, 1).toUpperCase().concat(ele.slice(1))).join(' ');
+    var quantityValue = Number(this.state.currentIngredientQuantity) // document.getElementById(qty).value;
     // best to get the measurement from the state
     // perhaps becuse inner html is defined before state is updated
     // var measurementValue = this.state.currentIngredientMeasurement;
     var measurementValue = document.getElementById(msr).value;
-
     if (ingredientValue === "") {
       window.alert("Enter an ingredient to add to product@@");
       return;
@@ -737,7 +776,7 @@ class SuggestProductForm extends Component {
 
     console.log(
       "current state of product index at Add Ingredient To product is : \n" +
-        this.state.currProductIndexInDBsProductsList
+      this.state.currProductIndexInDBsProductsList
     );
 
     // create state for each nutritional value
@@ -765,7 +804,7 @@ class SuggestProductForm extends Component {
     console.log("COMES IN addIngredientToProduct");
     event.preventDefault();
     var properIngredientStringSyntax;
-    var ingredientValue = document.getElementById("currentIngredient").value;
+    var ingredientValue = document.getElementById("currentIngredient").value.split(' ').map((ele) => ele.slice(0, 1).toUpperCase().concat(ele.slice(1))).join(' ')
     var quantityValue = document.getElementById(
       "currentIngredientQuantity"
     ).value;
@@ -806,7 +845,7 @@ class SuggestProductForm extends Component {
 
     console.log(
       "current state of product index at Add Ingredient To product is : \n" +
-        this.state.currProductIndexInDBsProductsList
+      this.state.currProductIndexInDBsProductsList
     );
 
     const searchResult = this.productNames.map(function callback(element) {
@@ -1003,7 +1042,7 @@ class SuggestProductForm extends Component {
     //-------------Submit remainder data of product to Mongo ------------------------------------------
     let suggestProductForm = new FormData();
     suggestProductForm.append("item_name", productName);
-
+    console.log(productImage1, productImage2, productImage3, productImage4, 'product image 1SßßS')
     if (productImage1) {
       suggestProductForm.append("item_images", productImage1);
     }
@@ -1017,6 +1056,14 @@ class SuggestProductForm extends Component {
       suggestProductForm.append("item_images", productImage4);
     }
 
+    if (!productImage1 && !productImage2 && !productImage3 && !productImage4) {
+      const img = await fetch(
+        "/assets/store_pics/no-image-product.png"
+      );
+      const blob = await img.blob()
+      suggestProductForm.append("item_images", blob);
+    }
+
     // suggestProductForm.append('product_images', productImage2);
     // suggestProductForm.append('product_images', productImage3);
     // suggestProductForm.append('product_images', productImage4);
@@ -1026,7 +1073,12 @@ class SuggestProductForm extends Component {
       console.log(individualDescriptions);
       // suggestProductForm.append('product_descriptions', individualDescriptions);
     });
-console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
+    console.log(
+      nutritionalStrings,
+      item_description,
+      ingredientStrings,
+      "pppr"
+    );
     const arr = nutritionalStrings.map((ele) => {
       let obj = {};
       let information = ele.split(":")[0]?.trim()?.toLowerCase();
@@ -1034,9 +1086,9 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
       information =
         information?.split(" ")?.length > 1
           ? information
-              ?.split(" ")
-              .map((ele) => ele.toLowerCase())
-              .join("_")
+            ?.split(" ")
+            .map((ele) => ele.toLowerCase())
+            .join("_")
           : information;
       obj[information] = value;
       return obj;
@@ -1058,7 +1110,7 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
       }
     }
 
-    productObject.product_size = sizeStrings.toString();
+    productObject.product_size = sizeStrings?.toString();
 
     let arr2 = [];
 
@@ -1067,6 +1119,7 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
     };
 
     for (let ele of Object.keys(description)) {
+      console.log(description[ele], description, 'description_')
       arr2.push({
         object_name: ele,
         object_quantity: description[ele].match(/\d+/)[0],
@@ -1078,31 +1131,25 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
       });
     }
 
-
-
     suggestProductForm.append("item_data", JSON.stringify(productObject));
 
     suggestProductForm.append("description", JSON.stringify(arr2));
 
     let arr_1 = [];
 
-    for(let ele of ingredientStrings){
+    for (let ele of ingredientStrings) {
       let obj = {};
-      const splited = ele.split(" ")
-      obj.item_name = splited.slice(3,)?.toString().split(',').join(' ')
-      obj.item_quantity = splited[0]
-      obj.item_measurement = splited[1]
+      const splited = ele.split(" ");
+      obj.item_name = splited.slice(3)?.toString().split(",").join(" ");
+      obj.item_quantity = splited[0];
+      obj.item_measurement = splited[1];
       obj.formatted_string_of_item = ele;
 
-      arr_1.push(obj)
+      arr_1.push(obj);
     }
 
-    suggestProductForm.append(
-      "formatted_ingredients",
-      JSON.stringify(arr_1)
-    );
-    console.log(ingredientStrings, 'nfnn')
-
+    suggestProductForm.append("formatted_ingredients", JSON.stringify(arr_1));
+    console.log(ingredientStrings, "nfnn");
 
     // suggestProductForm.append("store_available", '63d426b416b83177aaeaed96');
     suggestProductForm.append("item_type", this.props.suggestionType);
@@ -1140,13 +1187,11 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
     // var url = "http://localhost:5000/api/products/create/";
     var url = `${base_url}/items/`;
 
-    
-
     console.log("Printing Chunk Contents");
 
     var instructionData = JSON.parse(JSON.stringify(instructionGroupData));
-    
-    console.log(this.props.router, 'this.props.router');
+
+    console.log(this.props.router, "this.props.router");
     if (this.props.router?.query?.id) {
       url = url + `?action=update&_id=${this.props.router.query.id}`;
     }
@@ -1181,6 +1226,57 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
       });
   };
 
+  async preprocessImage() {
+    try {
+      this.setState({
+        ...this.state,
+        image: {
+          ...this.state.image,
+          loading: true
+        }
+      })
+      const form = new FormData()
+      this.state.image.file.map((entry) => form.append("image", entry))
+
+      const response = await axios.put("/items/content-extraction", form, {
+        headers: {
+          "Content-Type": "multipart/form-data"
+        }
+      })
+
+      const data = JSON.parse(response.data?.data || "{}")
+      this.setState({
+        ...this.state,
+        productName: data?.product_name || "",
+        productDescription: data?.product_intro || "",
+        sizeQuantity: data?.size?.size || "",
+        sizeMeasurement: data?.size?.msr || "",
+        sizeStrings: [`${data?.size?.size} ${data?.size?.msr}`],
+        ingredientStrings: data?.ingredients?.map((entry) => `${entry?.qty} ${entry?.msr} of ${entry?.name}`),
+        nutritionalStrings: data?.nutritional_info?.map((entry) => `${entry?.name} : ${entry?.qty}${entry?.msr}`),
+        image: {
+          ...this.state.image,
+          loading: false,
+          visible: false
+        },
+        productImage1: this.state.image.file[0],
+        productImagesData: [
+          ...this.state.productImagesData,
+          ...this.state.image.url,
+        ],
+      })
+    } catch (e) {
+      console.log(e)
+      this.setState({
+        ...this.state,
+        image: {
+          ...this.state.image,
+          loading: false
+        }
+      })
+    }
+  }
+
   deleteImages(id) {
     const deleteImg = this.state.productImagesData;
     deleteImg.splice(id, 1);
@@ -1189,9 +1285,121 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
   ///////////////////////////////////////////////////////////////////////////////////////
   render() {
     const { ingredientStrings, sizeStrings, nutritionalStrings } = this.state;
+    console.log(this.state.ingredientStrings, 'ingredientStrings')
 
     return (
       <div className={styles.suggestion_section_2}>
+        {
+          this.state.image.visible && <div
+            onClick={() => this.setState({
+              ...this.state,
+              image: {
+                ...this.state.image,
+                visible: false
+              }
+            })}
+            className="product_suggestion_popup_container">
+            <div
+              onClick={(e) => e.stopPropagation()}
+              className="product_suggestion_popup_child">
+              <div className="product_suggestion_popup_header">
+                <p>Upload your image</p>
+                <div
+                  onClick={() => this.setState({
+                    ...this.state,
+                    image: {
+                      ...this.state.image,
+                      visible: false
+                    }
+                  })}
+                >
+                  <IoIosCloseCircle size={25} />
+                </div>
+              </div>
+              <div className="product_suggestion_popup_body">
+                <p className="header_text">Upload your product image to scan (Jpg,png files only)</p>
+                <div className="product_suggestion_popup_box">
+
+                  {
+                    this.state.image.url.length ? this.state.image.url.map((entry) => (
+                      <div className="product_suggestion_popup_div">
+                        <div className="product_suggestion_popup_preview">
+                          {
+                            this.state.image.file && <>
+                              <div className="prod_box_1">
+                                <svg width="52" height="53" viewBox="0 0 52 53" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                  <path d="M50 4C51.1046 4 52 3.10457 52 2C52 0.895431 51.1046 1.19209e-07 50 1.19209e-07V4ZM0 4H50V1.19209e-07H0V4Z" fill="#00A001" />
+                                  <path d="M2.14186e-06 51C2.19014e-06 52.1046 0.895433 53 2 53C3.10457 53 4 52.1046 4 51L2.14186e-06 51ZM0 2L2.14186e-06 51L4 51L4 2L0 2Z" fill="#00A001" />
+                                </svg>
+                              </div>
+                              <div className="prod_box_2">
+                                <svg width="50" height="52" viewBox="0 0 50 52" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                  <path d="M2 4C0.895432 4 0 3.10457 0 2C0 0.895431 0.895432 1.19209e-07 2 1.19209e-07V4ZM50 4H2V1.19209e-07H50V4Z" fill="#00A001" />
+                                  <path d="M50 50C50 51.1046 49.1046 52 48 52C46.8954 52 46 51.1046 46 50L50 50ZM50 1L50 50L46 50L46 1L50 1Z" fill="#00A001" />
+                                </svg>
+                              </div>
+                              <div className="prod_box_3">
+                                <svg width="51" height="54" viewBox="0 0 51 54" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                  <path d="M49 50C50.1046 50 51 50.8954 51 52C51 53.1046 50.1046 54 49 54V50ZM0 50H49V54H0V50Z" fill="#00A001" />
+                                  <path d="M2.22928e-06 2C2.27756e-06 0.895431 0.895433 -4.82823e-08 2 0C3.10457 4.82823e-08 4 0.895432 4 2L2.22928e-06 2ZM0 53L2.22928e-06 2L4 2L4 53L0 53Z" fill="#00A001" />
+                                </svg>
+                              </div>
+                              <div className="prod_box_4">
+                                <svg width="50" height="53" viewBox="0 0 50 53" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                  <path d="M2 49C0.895432 49 0 49.8954 0 51C0 52.1046 0.895432 53 2 53V49ZM50 49H2V53H50V49Z" fill="#00A001" />
+                                  <path d="M50 2C50 0.895431 49.1046 -4.82823e-08 48 0C46.8954 4.82823e-08 46 0.895432 46 2L50 2ZM50 53L50 2L46 2L46 53L50 53Z" fill="#00A001" />
+                                </svg>
+                              </div>
+                            </>
+                          }
+                          {
+                            this.state.image.loading && <div className="scanner" />
+                          }
+                          {
+                            this.state.image.url.length && <div className="image-extraction-parent">
+                              <div
+                                style={{
+                                  backgroundImage: `url(${entry})`
+                                }}
+                                className="image-extraction" />
+                            </div>
+                          }
+                        </div>
+                      </div>
+                    )) : null
+                  }
+
+                  {
+                    this.state.image.url.length <= 3 ? <div onClick={() => this.ref.current?.click()} className="product_suggestion_popup_div">
+                      <svg width="32" height="23" viewBox="0 0 52 43" fill="none" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M11.2772 16.8281C5.957 18.0934 2 22.8766 2 28.584C2 35.2574 7.40989 40.6673 14.0833 40.6673C15.2281 40.6673 16.3357 40.5082 17.3851 40.2107" stroke="#F47900" stroke-width="4" stroke-linecap="round" stroke-linejoin="round" />
+                        <path d="M40.7329 16.8281C46.0532 18.0934 50.0101 22.8766 50.0101 28.584C50.0101 35.2574 44.6003 40.6673 37.9268 40.6673C36.782 40.6673 35.6744 40.5082 34.625 40.2107" stroke="#F47900" stroke-width="4" stroke-linecap="round" stroke-linejoin="round" />
+                        <path d="M40.667 16.5C40.667 8.49189 34.1751 2 26.167 2C18.1589 2 11.667 8.49189 11.667 16.5" stroke="#F47900" stroke-width="4" stroke-linecap="round" stroke-linejoin="round" />
+                        <path d="M17.7871 26.024L26.1663 17.6172L34.784 26.1678" stroke="#F47900" stroke-width="4" stroke-linecap="round" stroke-linejoin="round" />
+                        <path d="M26.167 38.2493V21.8906" stroke="#F47900" stroke-width="4" stroke-linecap="round" stroke-linejoin="round" />
+                      </svg>
+                      <p>Upload New Image</p>
+                    </div> : null
+                  }
+                </div>
+                <div className="product_suggestion_popup_btn">
+                  <button onClick={() => this.preprocessImage()}>Generate content</button>
+                </div>
+              </div>
+              <input onChange={(event) => {
+                this.setState({
+                  ...this.state,
+                  image: {
+                    ...this.state.image,
+                    url: [...this.state.image.url, URL.createObjectURL(event.target.files[0])],
+                    file: [...this.state.image.file, event.target.files[0]]
+                  }
+                })
+              }} className="img_ref" accept="image/*" ref={this.ref} type="file" />
+            </div>
+          </div>
+        }
+
         <form
           id="formproduct"
           className={styles.suggestion_forms}
@@ -1235,7 +1443,14 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
               <div className={styles.suggestion_form_image}>
                 <div className={styles.suggestion_form_image_col_1}>
                   <div
-                    onClick={() => this.uploadProductImage()}
+                    // onClick={() => this.uploadProductImage()}
+                    onClick={() => this.setState({
+                      ...this.state,
+                      image: {
+                        ...this.state.image,
+                        visible: true
+                      }
+                    })}
                     className={styles.suggestion_form_image_icon_con}
                   >
                     <AddIcon className={styles.suggestion_form_image_icon} />
@@ -1309,6 +1524,7 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
                     Quantity
                   </label>
                   <TextField
+                    inputProps={{ min: 0 }}
                     fullWidth
                     id="sizeQuantity"
                     type="number"
@@ -1330,7 +1546,7 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
                   </label>
                   <Autocomplete
                     id="sizeMeasurement"
-                    options={this.measurements.map((option) => option)}
+                    options={this.state.measurements.map((option) => option)}
                     value={this.state.sizeMeasurement}
                     x
                     onChange={this.handleSizeMeasurement}
@@ -1382,7 +1598,7 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
               </label>
               <Autocomplete
                 id="currentIngredient"
-                options={this.productNames.map((option) => option)}
+                options={this.props.productNames.map((option) => option)}
                 // onChange={(ev)=>this.onTextFieldChange(ev)}
                 value={this.state.currentIngredient}
                 onChange={(ev, val) => this.handleProductNameInput(ev, val)}
@@ -1399,7 +1615,13 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
                 fullWidth
               />
             </div>
-            <div className={styles.suggestion_form_2_col}>
+            <div className={styles.suggestion_form_2_col} onKeyDown={(event) => {
+              // Check if the pressed key is Enter
+              if (event.key === "Enter") {
+                event.preventDefault(); // Prevent default form submission
+                this.addIngredientToProduct(event);
+              }
+            }}>
               <div className={styles.suggestion_form_2_col_1}>
                 <div className={styles.suggestion_form_group}>
                   <label
@@ -1409,13 +1631,14 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
                     Quantity
                   </label>
                   <TextField
+                    inputProps={{ min: 0 }}
                     fullWidth
                     id="currentIngredientQuantity"
                     type="number"
                     onChange={this.onTextFieldChange}
                     variant="outlined"
                     placeholder="1.."
-                    // value={this.state.currentIngredientQuantity}
+                  // value={this.state.currentIngredientQuantity}
                   />
                 </div>
               </div>
@@ -1464,6 +1687,7 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
                 <Chip
                   key={index}
                   label={data}
+
                   className={styles.chip}
                   onClick={() => this.handleDeleteIngredientChip(data)}
                   onDelete={() => this.handleDeleteIngredientChip(data)}
@@ -1472,7 +1696,17 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
             </Stack>
           </div>
           <h3>Product Description</h3>
-          <div className={styles.suggestion_form}>
+          <div className={styles.suggestion_form} onKeyDown={(event) => {
+            if (event.key === "Enter") {
+              event.preventDefault();
+              this.addNutritionalFactToProduct(
+                event,
+                "currentNutritionName",
+                "currentIngredientQuantity",
+                "currentIngredientMeasurement1"
+              )
+            }
+          }}>
             <div className={styles.suggestion_form_group}>
               <label
                 htmlFor="currentIngredient"
@@ -1511,13 +1745,14 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
                     Quantity
                   </label>
                   <TextField
+                    inputProps={{ min: 0 }}
                     fullWidth
                     id="currentIngredientQuantity"
                     type="number"
                     onChange={this.onTextFieldChange}
                     variant="outlined"
                     placeholder="1.."
-                    // value={this.state.currentIngredientQuantity}
+                    value={this.state.currentIngredientQuantity}
                   />
                 </div>
               </div>
@@ -1532,7 +1767,7 @@ console.log(nutritionalStrings, item_description, ingredientStrings, 'pppr')
                   </label>
                   <Autocomplete
                     id="currentIngredientMeasurement1"
-                    options={this.measurements.map((option) => option)}
+                    options={this.state.measurements.map((option) => option)}
                     // value={this.state.currentIngredientMeasurement}
                     onChange={this.handleIngredientMeasurement}
                     freeSolo
