@@ -1,30 +1,51 @@
-export const FETCH_START = 'fetch_start';
-export const FETCH_SUCCESS = 'fetch_success';
-export const FETCH_ERROR = 'fetch_error';
-export const GET_PATH = 'get_path';
-export const TRIGGER_SNACK = 'trigger_snack';
+export const FETCH_START = "fetch_start";
+export const FETCH_SUCCESS = "fetch_success";
+export const FETCH_ERROR = "fetch_error";
+export const GET_PATH = "get_path";
+export const TRIGGER_SNACK = "trigger_snack";
+
+export const SIGNIN_USER = "signin_user";
+export const SIGNIN_USER_SUCCESS = "signin_user_success";
+export const SIGNOUT_USER = "signout_user";
+export const SIGNOUT_USER_SUCCESS = "signout_user_success";
+export const INIT_URL = "init_url";
+export const USER_DATA = "user_data";
+export const USER_TOKEN_SET = "user_token_set";
+export const USER_ROLE = "user_role_set";
+export const CUSTOMER_ID = "customer_id";
+export const IS_AUTHENTICATED = "isAuthenticated";
+export const IS_VERIFIED = "isVerified";
+export const EMAIL_VERIFIED = 'isEmailVerified';
+export const PHONE_NUMBER_VERIFIED = 'isNumberVerified';
+export const OPEN_LOGIN = "openLogin";
+
+export const SEARCH_TITLE = "search_title";
+export const CART_LIST = "cart_list";
+export const TRENDING_VIDEOS = "trending_videos";
+export const WATCHING_VIDEOS = "watching_videos";
+
+export const CATEGORY_LIST = "category_list";
+
+export const JAWBONE_HANDLE = "jawbone_handle";
+
+export const USER_ADDED_TO_STORE = "user_added_to_store";
+export const USER_REMOVE_FROM_STORE = "user_remove_from_store";
+
+//CART CONSTANT ADDED BY PETERS - 12/6/2024
+export const CART_ADD_ITEM = 'cart_add_and_update_item';
+export const CART_REMOVE_ITEM = 'cart_remove_item';
+export const CART_HAS_ITEM = 'cart_has_item';
+export const CART_EMPTY = 'empty_cart';
+export const CART_DELETE_ITEM = 'delete_cart_items';
+export const FETCH_CART = "fetch_cart";
+export const CART_STATE = "cart_state";
 
 
-export const SIGNIN_USER = 'signin_user';
-export const SIGNIN_USER_SUCCESS = 'signin_user_success';
-export const SIGNOUT_USER = 'signout_user';
-export const SIGNOUT_USER_SUCCESS = 'signout_user_success';
-export const INIT_URL = 'init_url';
-export const USER_DATA = 'user_data';
-export const USER_TOKEN_SET = 'user_token_set';
-export const USER_ROLE = 'user_role_set';
-export const CUSTOMER_ID = 'customer_id';
-export const IS_AUTHENTICATED = 'isAuthenticated';
-export const IS_VERIFIED = 'isVerified';
-export const OPEN_LOGIN = 'openLogin';
-
-export const SEARCH_TITLE = 'search_title';
-export const CART_LIST = 'cart_list';
-export const TRENDING_VIDEOS = 'trending_videos';
-export const WATCHING_VIDEOS = 'watching_videos';
-
-export const CATEGORY_LIST = 'category_list';
-
-export const JAWBONE_HANDLE = 'jawbone_handle';
+//Orders
+export const CREATE_ORDER = 'create_order';
+export const ORDER_REMOVE_ITEM = 'remove_order';
+export const ORDER_DELETE = 'delete_orders';
 
 
+export const  OPEN_VERIFICATION = 'open_verification';
+export const  LOGIN_ON_VERIFICATION = 'login_on_verification';
