@@ -13,7 +13,7 @@ const proxy = require("http-proxy-middleware");
 // };
 module.exports = function (app) {
   app.use(proxy("/api/", {
-    target: "https://chopchowserver.vercel.app/api/",
+    target:  "https://chopchowserver.vercel.app/api/", //"https://localhost:5000",
     secure: false
   }));
 };
